% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols_result_df_pd_changes.R
\name{add_cols_result_df_pd_changes}
\alias{add_cols_result_df_pd_changes}
\title{Function that renames Merton model baseline result columns and adds additional
columns for the second application of the model in the late & sudden scenario.}
\usage{
add_cols_result_df_pd_changes(data, horizon = NULL)
}
\arguments{
\item{data}{A dataframe that contains the result data from the baseline
Merton model}

\item{horizon}{A cahracter vector of length 1 indicating whether to return
a result obejct for overall or yearly results. Must be "annual" or "overall"}
}
\description{
Function that renames Merton model baseline result columns and adds additional
columns for the second application of the model in the late & sudden scenario.
}
\seealso{
Other helper functions: 
\code{\link{create_empty_result_df_pd_changes}()}
}
\concept{helper functions}
