% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_price_data}
\alias{add_price_data}
\title{Augment pacta and fin data with price data}
\usage{
add_price_data(pacta_and_fin, price_data)
}
\arguments{
\item{pacta_and_fin}{Tibble holding pacta results and fin data.}

\item{price_data}{Tibble holding price data or NULL for}
}
\value{
Tibble \code{pacta_and_fin} holding additionally price_data
}
\description{
Joins \code{price_data} to \code{pacta_and_fin} and adds artificial price_data (1) for
automotive sector, for which to data no price data are available.
}
