% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_results.R
\name{aggregate_results}
\alias{aggregate_results}
\title{Aggregate results}
\usage{
aggregate_results(results_list, sensitivity_analysis_vars)
}
\arguments{
\item{results_list}{A list of results.}

\item{sensitivity_analysis_vars}{String vector holding names of iteration
arguments.}
}
\value{
A list including basic and aggregated results.
}
\description{
Function aggregates results to the expected levels. For market risk, the most
granular level company-technology value changes are aggregated to (1) the
company level and then (2) the portfolio level with splits by technology,
sector, all analyzed sectors and overall portfolio impact. For the credit
risk results, the most granular level already starts at company level. Here,
we additionally aggregate to the portfolio level, providing a sector split of
the credit risk impact on the portfolio. In all cases, the aggregation to the
portfolio level is done weighting the holdings in the given companies by
exposure. List element entry \code{results} is split into market risk and credit
risk results for company and portfolio level.
}
