% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_pd_change_technology_shock_year.R
\name{annual_pd_change_sector_shock_year}
\alias{annual_pd_change_sector_shock_year}
\title{Display the annual pd changes per shock year and technology.}
\usage{
annual_pd_change_sector_shock_year(
  data,
  shock_year_filter = NULL,
  geography_filter = NULL
)
}
\arguments{
\item{data}{A data.frame that contains Merton output data.}

\item{shock_year_filter}{Numeric. Vector that specifies the respective start
years of different policy shocks.}

\item{geography_filter}{Character. A vector containing the geographical
specification of the scenario}
}
\value{
ggplot object
}
\description{
Display the annual pd changes per shock year and technology.
}
\seealso{
Other qa graphing functions: 
\code{\link{annual_pd_change_company_sector}()},
\code{\link{overall_pd_change_company_sector}()},
\code{\link{overall_pd_change_sector_shock_year}()},
\code{\link{show_carbon_budget}()},
\code{\link{show_impact_by_shock_year}()},
\code{\link{show_pf_technology_shares}()},
\code{\link{show_price_trajectories}()},
\code{\link{show_prod_baseline_target_ls_pf}()},
\code{\link{show_prod_trajectories}()},
\code{\link{show_var_change_by_shock_year}()}
}
\concept{qa graphing functions}
