% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_without_duplicate_cols}
\alias{append_without_duplicate_cols}
\title{Append without duplicat cols}
\usage{
append_without_duplicate_cols(data_x, data_y)
}
\arguments{
\item{data_x}{A tibble.}

\item{data_y}{A tibble.}
}
\value{
A tibble
}
\description{
Function binds \code{data_x} and \code{data_y} together. In case of overlapping cols,
cols from \code{data_y} are used.
}
