% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_filters.R
\name{apply_filters}
\alias{apply_filters}
\title{Filter input company or portfolio data to match model specifications and
parameters. Companies with zero production in the start year are removed as
they will lead to errors in calculations down the line.}
\usage{
apply_filters(
  data,
  investor = NULL,
  sectors = NULL,
  technologies = NULL,
  scenario_geography_filter = NULL,
  scenarios = NULL,
  allocation_method = NULL,
  start_analysis = NULL
)
}
\arguments{
\item{data}{A dataframe containing the production forecasts of companies
(in the portfolio). Usually a PACTA output.}

\item{investor}{Character. A vector of length 1 containing the name of the
investor for whom to run the analysis.}

\item{sectors}{Character. A vector containing the sectors to include in the
analysis. In most cases this will be a subset of PACTA sectors.}

\item{technologies}{Character. A vector containing the technologies to
include in the analysis. In most cases this will be a subset of PACTA
technologies.}

\item{scenario_geography_filter}{Character. A vector of length 1 that
indicates which geographic scenario to apply in the analysis.}

\item{scenarios}{Character. A vector containing the unique names of the
scenarios to include in the analysis. This refers to baseline, target and
aligned scenarios.}

\item{allocation_method}{Character. A vector of length 1 indicating the
set of PACTA data to be used in the analysis, based on the choice of an
allocation rule.}

\item{start_analysis}{Numeric. A vector of length 1 indicating the start
year of the analysis.}
}
\description{
Filter input company or portfolio data to match model specifications and
parameters. Companies with zero production in the start year are removed as
they will lead to errors in calculations down the line.
}
