% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_flat_multiplier}
\alias{assign_flat_multiplier}
\title{Assign value of flat multiplier}
\usage{
assign_flat_multiplier(asset_type)
}
\arguments{
\item{asset_type}{String holding asset_type. For accepted values compare
\code{stress_test_arguments}.}
}
\value{
A double holding value of the flat multiplier.
}
\description{
Assign value of flat multiplier based on \code{asset_type}.
}
