% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_lgd}
\alias{assign_lgd}
\title{Assign value of lgd}
\usage{
assign_lgd(asset_type, lgd_senior_claims, lgd_subordinated_claims)
}
\arguments{
\item{asset_type}{String holding asset_type, for allowed value compare
\code{asset_types_lookup}.}

\item{lgd_senior_claims}{Numeric, holding the loss given default for senior
claims, for accepted value range check \code{stress_test_arguments}.}

\item{lgd_subordinated_claims}{Numeric, holding the loss given default for
subordinated claims, for accepted value range check
\code{stress_test_arguments}.}
}
\value{
A numeric holding value of lgd.
}
\description{
Assigns value of lgd based on \code{asset_type}. Can be from \code{lgd_senior_claims}
or \code{lgd_subordinated_claims}.
}
