% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_late_sudden_prices.R
\name{calc_late_sudden_prices}
\alias{calc_late_sudden_prices}
\title{Calculate late and sudden prices}
\usage{
calc_late_sudden_prices(
  price_data,
  baseline_scenario,
  transition_scenario,
  start_year
)
}
\arguments{
\item{price_data}{A tibble holding price data.}

\item{baseline_scenario}{String holding name of the baseline scenario.}

\item{transition_scenario}{Tibble with 1 row holding at least variables
\code{year_of_shock} and \code{duration_of_shock}.}

\item{start_year}{Numeric, holding start year of analysis.}
}
\value{
A tibble holding late_and_sudden_prices
}
\description{
Wrapper function around call to late_sudden_prices.
}
