% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_annual_net_profits.R
\name{calculate_annual_net_profits}
\alias{calculate_annual_net_profits}
\title{Calculate annual net profits under baseline and late & sudden scenarios.}
\usage{
calculate_annual_net_profits(
  data,
  carbon_price_data = NULL,
  baseline_tax = NULL,
  late_sudden_tax = NULL
)
}
\arguments{
\item{data}{A dataframe containing the annual production and price data}

\item{carbon_price_data}{A dataframe containing the annual development of
carbon prices under different scenarios.}

\item{baseline_tax}{A character vector of length 1, specifying the baseline
scenario for the carbon tax.}

\item{late_sudden_tax}{A character vector of length 1, specifying the late &
sudden scenario for the carbon tax.}
}
\description{
This function calculates the annual net profits for the baseline
and the late & sudden shock scenarios for each company-technology combination.
It is based on the production trajectories, market prices (IEA by default),
company net profit margins (regional/sector averages, where company level
not known), forward looking emission factors and a carbon tax (defaults to
NGFS carbon prices). We model the the production channel and the price
channel. More details can be obtained in the general documentation.
}
