% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_annual_profits}
\alias{calculate_annual_profits}
\title{Calculate annual profits}
\usage{
calculate_annual_profits(
  data,
  baseline_scenario,
  shock_scenario,
  end_year,
  discount_rate,
  growth_rate,
  log_path
)
}
\arguments{
\item{data}{data frame containing the full trajectory company data}

\item{baseline_scenario}{Character. A string that indicates which
of the scenarios included in the analysis should be used to set the
baseline technology trajectories.}

\item{shock_scenario}{Character. A string that indicates which
of the scenarios included in the analysis should be used to set the
late & sudden technology trajectories.}

\item{end_year}{Numeric, holding end year of analysis.}

\item{discount_rate}{Numeric, that holds the discount rate of dividends per
year in the DCF. For accepted range compare \code{stress_test_arguments}.}

\item{growth_rate}{Numeric, that holds the terminal growth rate of profits
beyond the \code{end_year} in the DCF.}

\item{log_path}{String holding path to log file.}
}
\value{
A tibble holding annual profits
}
\description{
Wrapper function to calculate discounted annual profits and terminal value.
}
