% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_aum.R
\name{calculate_aum}
\alias{calculate_aum}
\title{Calculate aum}
\usage{
calculate_aum(sector_exposures)
}
\arguments{
\item{sector_exposures}{A tibble holding sector exposure.}
}
\value{
A tibble holding columns investor_name, portfolio_name and
asset_portfolio_value.
}
\description{
Function calculated asset under management for portfolio from sector
exposures.
}
