% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_exposure_by_technology_and_company}
\alias{calculate_exposure_by_technology_and_company}
\title{Calculate exposure_by_technology_and_company}
\usage{
calculate_exposure_by_technology_and_company(
  asset_type,
  input_data_list,
  start_year,
  time_horizon,
  scenario_to_follow_shock,
  log_path
)
}
\arguments{
\item{asset_type}{String holding asset_type. For accepted values compare
\code{stress_test_arguments}.}

\item{log_path}{String holding path to log file.}
}
\value{
A tibble holding exposure_by_technology_and_company.
}
\description{
Wrapper to calculate exposure_by_technology_and_company.
}
