% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_functions.R
\name{calculate_net_profits}
\alias{calculate_net_profits}
\title{Calculates annual net profits on the company-technology level for the
baseline and late and sudden scenarios. Climate laggards which need to build
out their production in increasing technologies to compensate for their
missed targets, are "punished" by adjusting the net profit margin on their
additional build out based on their proximity to target and market share within
the given technology. This ensures that late build out will not
proportionally translate into increased profits. The adjustment for the
market share of the company accounts for the fact that companies of different
size will face different degrees of difficulty catching up.}
\usage{
calculate_net_profits(data)
}
\arguments{
\item{data}{A data frame containing the production forecasts of companies
under baseline and late and sudden, market prices/costs, company net profit
margins, the proximity to target in the production forecast period and an
indication of the direction of the technology.}
}
\description{
Calculates annual net profits on the company-technology level for the
baseline and late and sudden scenarios. Climate laggards which need to build
out their production in increasing technologies to compensate for their
missed targets, are "punished" by adjusting the net profit margin on their
additional build out based on their proximity to target and market share within
the given technology. This ensures that late build out will not
proportionally translate into increased profits. The adjustment for the
market share of the company accounts for the fact that companies of different
size will face different degrees of difficulty catching up.
}
