% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_functions.R
\name{calculate_net_profits_shock_declining_technologies_carbon_tax}
\alias{calculate_net_profits_shock_declining_technologies_carbon_tax}
\title{Calculates annual net profits on the company-technology level for the
baseline and late and sudden scenarios - with a carbon tax being added.}
\usage{
calculate_net_profits_shock_declining_technologies_carbon_tax(
  data,
  shock_year,
  carbon_data
)
}
\arguments{
\item{data}{A data frame containing the production forecasts of companies
under baseline and late and sudden, market prices/costs, company net profit
margins, the proximity to target in the production forecast period and an
indication of the direction of the technology.}

\item{shock_year}{A numeric vector of length one that indicates in which year
the policy shock strikes in a given scenario.}

\item{carbon_data}{NGFS carbon prices.}
}
\description{
Calculates annual net profits on the company-technology level for the
baseline and late and sudden scenarios - with a carbon tax being added.
}
