% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_annual_pd_changes.R
\name{calculate_pd_change_annual}
\alias{calculate_pd_change_annual}
\title{Calculate change in probabilities of default (PDs) of loans connected to
companies at hand. This is based on the equity values derived from the DCF
model. Said Equity values are used as different starting points for the
Merton model (one reflecting the business as usual baseline scenario, the
other reflecting the late & sudden shock scenario). Since we calculate the PDs
on an annual basis (start year up to year t), we use different NPVs based on
the DCF model for every year that we calculate a PD for. Only discounted
profits up until the year in question will be taken into account as the equity
starting points for each of the calculations. Prior to a policy shock, this
implies equal starting points for baseline and late and sudden scenarios and
only after the shock, we get diverging starting points.}
\usage{
calculate_pd_change_annual(
  data,
  shock_year = NULL,
  end_of_analysis = NULL,
  exclusion = NULL
)
}
\arguments{
\item{data}{A dataframe containing the (discounted) annual profits}

\item{shock_year}{A numeric vector of length one that indicates in which year
the policy shock strikes in a given scenario}

\item{end_of_analysis}{A numeric vector of length one that indicates until
which year the analysis runs}

\item{exclusion}{Optional. A dataframe with two character columns, "company_name" and
"technology", that lists which technologies from which companies should be
set to 0 in the remainder of the analysis.}
}
\description{
Calculate change in probabilities of default (PDs) of loans connected to
companies at hand. This is based on the equity values derived from the DCF
model. Said Equity values are used as different starting points for the
Merton model (one reflecting the business as usual baseline scenario, the
other reflecting the late & sudden shock scenario). Since we calculate the PDs
on an annual basis (start year up to year t), we use different NPVs based on
the DCF model for every year that we calculate a PD for. Only discounted
profits up until the year in question will be taken into account as the equity
starting points for each of the calculations. Prior to a policy shock, this
implies equal starting points for baseline and late and sudden scenarios and
only after the shock, we get diverging starting points.
}
