% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_filter_data.R
\name{check_and_filter_data}
\alias{check_and_filter_data}
\title{Check and filter data}
\usage{
check_and_filter_data(
  st_data_list,
  start_year,
  end_year,
  scenarios_filter,
  scenario_geography_filter
)
}
\arguments{
\item{st_data_list}{A list holding imported and prewrangled stress test input
data.}

\item{start_year}{Numeric holding start year of analysis.}

\item{end_year}{Numeric holding end year of analysis.}

\item{scenarios_filter}{String vector holding name of baseline and shock
scenario.}

\item{scenario_geography_filter}{String holding name of considered
geographical regions.}
}
\value{
List \code{st_data_list} with tibbles subset to rows required for
analysis.
}
\description{
Function filters dataset so that only rows relevant for analysis remain. Also
an error is thrown if duplicate entries are detected.
}
\details{
Note that dataset \code{sector_exposures} is not filtered for considered sectors
as data is used in \code{\link[=calculate_aum]{calculate_aum()}} where the asset under management of the
entire portfolio is calculated.
}
