% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_company_ticker_mapping}
\alias{check_company_ticker_mapping}
\title{Check mapping of companies to corporate bond tickers}
\usage{
check_company_ticker_mapping(data)
}
\arguments{
\item{data}{A tibble holding at least columns \code{company_name} and
\code{corporate_bond_ticker}.}
}
\value{
Returns data invisibly
}
\description{
There needs to be an n x 1 mapping between companies and
corporate bond tickers. Function verifies this is the case by checking that
companies do not map to > 1 corporate bond ticker.
}
