% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_expected_missings}
\alias{check_expected_missings}
\title{Check expected missing patterns}
\usage{
check_expected_missings(data)
}
\arguments{
\item{data}{Data that expected missings are checked on.}
}
\value{
Returns input invisibly.
}
\description{
On certain variables missings appear that are not problematic but are
expected. Here it is checked whether those missings adher to expected
patterns.
}
