% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_and_check.R
\name{check_financial_data}
\alias{check_financial_data}
\title{Check financial data}
\usage{
check_financial_data(financial_data, asset_type)
}
\arguments{
\item{financial_data}{A data set of \code{financial_data}.}

\item{asset_type}{A string indicating if company data are for analysis for
bond or equity.}
}
\value{
A prewrangled \code{financial_data} set.
}
\description{
Applies sanity checks to financial data. Also remove column
corporate_bond_ticker if \code{asset_type} is not bonds.
}
