% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_level_availability}
\alias{check_level_availability}
\title{Check that required levels of a variable are available in data}
\usage{
check_level_availability(data, expected_levels_list)
}
\arguments{
\item{data}{A tibble.}

\item{expected_levels_list}{List holding variable names along with their
expected levels. Variables must be present in data.}
}
\value{
Returns \code{data} invisibly.
}
\description{
Check that required levels of a variable are available in data
}
