% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_and_check.R
\name{check_results}
\alias{check_results}
\title{Check results}
\usage{
check_results(wrangled_results_list, sensitivity_analysis_vars)
}
\arguments{
\item{wrangled_results_list}{A list of wrangled results.}

\item{sensitivity_analysis_vars}{String vector holding names of iteration
arguments.}
}
\value{
\code{wrangled_results_list}
}
\description{
Function checks results for missings and duplicates.
}
