% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_results_structure}
\alias{check_results_structure}
\title{Checks structure of results}
\usage{
check_results_structure(data, name_data, cuc_cols)
}
\arguments{
\item{data}{Tibble holding results.}

\item{name_data}{Name of results data.}

\item{cuc_cols}{Vector of cols the combination of which needs to be unique.}
}
\value{
Tibble \code{data}.
}
\description{
Wrapper to call \code{\link[=report_all_duplicate_kinds]{report_all_duplicate_kinds()}} and \code{\link[=report_missings]{report_missings()}} on
results.
}
