% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_sector_tech_mapping}
\alias{check_sector_tech_mapping}
\title{Check that sector to ald_business_unit mapping corresponds to template}
\usage{
check_sector_tech_mapping(
  data,
  sector_col = "ald_sector",
  mapper_template = p4i_p4b_sector_technology_lookup
)
}
\arguments{
\item{data}{A tibble.}

\item{sector_col}{Name of column holding sector (needed due to naming
inconsistencies.)}

\item{mapper_template}{A tibble holding, among others, mapping of sectors to
technologies.}
}
\value{
Returns \code{data} invisibly.
}
\description{
Check that sector to ald_business_unit mapping corresponds to template
}
