% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_technology_availability}
\alias{check_technology_availability}
\title{Check that required technologies are provided}
\usage{
check_technology_availability(data, expected_technologies)
}
\arguments{
\item{data}{A tibble holding at least column \code{technology}}

\item{expected_technologies}{A vector holding names of expected technologies.}
}
\value{
Returns \code{data} invisibly.
}
\description{
Check that required technologies are provided
}
