% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_and_check.R
\name{check_valid_financial_data_values}
\alias{check_valid_financial_data_values}
\title{Check if values in financial data are plausible}
\usage{
check_valid_financial_data_values(financial_data)
}
\arguments{
\item{financial_data}{A data set of \code{financial_data}.}
}
\description{
Checks that numeric columns hold values in acceptable ranges.
}
