% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_asset_value_at_risk.R
\name{company_asset_value_at_risk}
\alias{company_asset_value_at_risk}
\title{Calculate percentage value change between scenarios for equity (and
temporarily other asset types) on the company-technology level}
\usage{
company_asset_value_at_risk(
  data,
  terminal_value = NULL,
  shock_scenario = NULL,
  div_netprofit_prop_coef = NULL,
  plan_carsten = NULL,
  port_aum = NULL,
  flat_multiplier = NULL
)
}
\arguments{
\item{data}{A dataframe containing the (discounted) annual profits}

\item{terminal_value}{Numeric. A ratio to determine the share of the
discounted value used in the terminal value calculation beyond the
projected time frame.}

\item{shock_scenario}{A dataframe containing the specification of the
shock scenario at hand}

\item{div_netprofit_prop_coef}{Numeric. A coefficient that determines how
strongly the future dividends propagate to the company value}

\item{plan_carsten}{A dataframe that contains the share of the portfolio
value of each company-technology combination. Used to quantify the impact of
the company-tech level shock on higher levels of aggregation in the pf}

\item{port_aum}{A dataframe containing the value of the portfolio for
the asset type at hand}

\item{flat_multiplier}{Numeric. A ratio that determines for the asset type
if how strongly the DCF should propagate to value changes.}
}
\description{
Calculate percentage value change between scenarios for equity (and
temporarily other asset types) on the company-technology level
}
