% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_expected_loss.R
\name{company_expected_loss}
\alias{company_expected_loss}
\title{Calculate percentage value change between scenarios for equity (and
temporarily other asset types) on the company-technology level}
\usage{
company_expected_loss(
  data,
  loss_given_default = NULL,
  exposure_at_default = NULL,
  port_aum = NULL
)
}
\arguments{
\item{data}{A dataframe containing the (discounted) annual profits}

\item{loss_given_default}{A dataframe containing sectoral percentages of how
large the share of lost value in a loan is, in case of default.}

\item{exposure_at_default}{A dataframe that contains the share of the
portfolio value of each company-technology combination. Used to quantify
the impact of the company-tech level shock on higher levels of aggregation
in the portfolio}

\item{port_aum}{A dataframe containing the value of the portfolio for
the asset type at hand}
}
\description{
Calculate percentage value change between scenarios for equity (and
temporarily other asset types) on the company-technology level
}
