% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cap_to_generation.R
\name{convert_power_cap_to_generation}
\alias{convert_power_cap_to_generation}
\title{Translate power capacity to power generation}
\usage{
convert_power_cap_to_generation(data, capacity_factors_power = NULL)
}
\arguments{
\item{data}{A data frame filtered and wrangled company level production
forecasts (of the companies in the portfolio). Usually based on PACTA
output.}

\item{capacity_factors_power}{A data frame containing capacity factors to
translate company level power capacity to units sold. Contains information
on the technology (pwoer sector only) and scenario_geography levels.}
}
\description{
Units of generated power are assumed to be sold and hence get priced in the
net profit calculations. This also entails converting MWh into MW per year,
since we calculate yearly profits. Note: For use in webscripts
\code{\link[=convert_cap_to_generation]{convert_cap_to_generation()}} is used currently, which only distinguishes
capacity factor by technology and scenario_geography, whereas this function
distinguishes further by year and scenario.
}
