% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_automotive_scenario.R
\name{correct_automotive_scendata}
\alias{correct_automotive_scendata}
\title{interpolate automotive sector scenario data for a specified set of years
note: this function should eventually be moved to the data_prep repository}
\usage{
correct_automotive_scendata(data, interpolation_years = NULL)
}
\arguments{
\item{data}{Data frame. Use the data generated by loading and basic wrangling of the scenario data}

\item{interpolation_years}{Numeric vector. Vector of years which ought to be replaced by the interpolation}
}
\value{
Data frame
}
\description{
interpolate automotive sector scenario data for a specified set of years
note: this function should eventually be moved to the data_prep repository
}
