% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_empty_result_df_pd_changes.R
\name{create_empty_result_df_pd_changes}
\alias{create_empty_result_df_pd_changes}
\title{Function that creates an empty tibble/data.frame for results of the Merton
model.}
\usage{
create_empty_result_df_pd_changes(data, horizon = NULL)
}
\arguments{
\item{data}{A dataframe that contains the input data to the Merton model}

\item{horizon}{A cahracter vector of length 1 indicating whether to return
a result obejct for overall or yearly results. Must be "annual" or "overall"}
}
\description{
Function that creates an empty tibble/data.frame for results of the Merton
model.
}
\seealso{
Other helper functions: 
\code{\link{add_cols_result_df_pd_changes}()}
}
\concept{helper functions}
