% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_stressdata_masterdata_file_paths}
\alias{create_stressdata_masterdata_file_paths}
\title{Get path for stress test masterdata files}
\usage{
create_stressdata_masterdata_file_paths(data_prep_timestamp, twodii_internal)
}
\arguments{
\item{data_prep_timestamp}{Character scalar holding timestamp of data prep.}

\item{twodii_internal}{Boolean, if TRUE 2dii internal mode is used.}
}
\value{
A list with 2 file paths.
}
\description{
Get path for stress test masterdata files when called in 2dii internal mode.
}
