% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{customise_output_path}
\alias{customise_output_path}
\title{Customise output path}
\usage{
customise_output_path(output_path)
}
\arguments{
\item{output_path}{String holding path to st output folder.}
}
\value{
Path to subdirectory in st output folder.
}
\description{
Checks for existence of provided \code{output_path} and extends it to hold a
subdirectory names after the timestamp.
}
