% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{degrees}
\alias{degrees}
\title{Insert the symbol for degrees}
\usage{
degrees()
}
\value{
A string.
}
\description{
Insert the symbol for degrees
}
\examples{
degrees()
glue::glue("2{degrees()} Investing Initiative")
}
\seealso{
Other miscellaneous utility functions: 
\code{\link{data_path}()},
\code{\link{dropbox_exists}()},
\code{\link{get_st_data_path}()}
}
\concept{miscellaneous utility functions}
