% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_functions.R
\name{dividend_discount_model}
\alias{dividend_discount_model}
\title{Calculates discounted net profits based on a dividends discount model}
\usage{
dividend_discount_model(data, discount_rate)
}
\arguments{
\item{data}{A data frame containing the annual net profits on company
ald_business_unit level}

\item{discount_rate}{Numeric, that holds the discount rate of dividends per
year in the DCF.}
}
\description{
Calculates discounted net profits based on a dividends discount model
}
