% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_companies.R
\name{exclude_companies}
\alias{exclude_companies}
\title{Effectively exclude certain technologies of companies from calculations}
\usage{
exclude_companies(
  data,
  exclusion = NULL,
  scenario_baseline = NULL,
  scenario_ls = NULL
)
}
\arguments{
\item{data}{A dataframe of the annual profits by company-tech for one asset
type.}

\item{exclusion}{A dataframe with two character columns, "company_name" and
"technology", that lists which technologies from which companies should be
set to 0 in the reaminder of the analysis.}

\item{scenario_baseline}{Character. String that defines which scenario is
considered baseline. Usually this should be the same as throughout the
workflow.}

\item{scenario_ls}{Character. String that defines which scenario is followed
in late and sudden.  Usually this should be the same as throughout the
workflow.}
}
\description{
NOTE: If \code{exclusion} is NULL \code{data} are returned.
}
\concept{processing functions}
