% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_scenario_trajectory.R
\name{extend_scenario_trajectory}
\alias{extend_scenario_trajectory}
\title{Extend the scenario pathways based on the fair share approach (now known as
market share approach).  We first ensure that all scenarios are extended to
the same point in time (start of the analysis plus forecast time frame.
Then we take the first value of the production on company level as the start
value and from the second year onward, we calculate the level of production
for a company under the scenarios at hand by applying the relative changes
for the respective years to the absolute values of the previous year.}
\usage{
extend_scenario_trajectory(
  data,
  scenario_data = NULL,
  start_analysis = NULL,
  end_analysis = NULL,
  time_frame = NULL
)
}
\arguments{
\item{data}{A data frame containing the production forecasts of companies
(in the portfolio). Pre-processed to fit analysis parameters and after
conversion of power capacity to generation.}

\item{scenario_data}{A data frame containing scenario data for the specified
parameters of the analysis, including the business as usual and target
scenarios, the relevant scenario geography and time frame for each of the
technologies.}

\item{start_analysis}{Numeric. A vector of length 1 indicating the start
year of the analysis.}

\item{end_analysis}{Numeric. A vector of length 1 indicating the end
year of the analysis.}

\item{time_frame}{Numeric. A vector of length 1 indicating the number of years
for which forward looking production data is considered.}
}
\description{
Extend the scenario pathways based on the fair share approach (now known as
market share approach).  We first ensure that all scenarios are extended to
the same point in time (start of the analysis plus forecast time frame.
Then we take the first value of the production on company level as the start
value and from the second year onward, we calculate the level of production
for a company under the scenarios at hand by applying the relative changes
for the respective years to the absolute values of the previous year.
}
