% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fail_if_input_args_are_missing}
\alias{fail_if_input_args_are_missing}
\title{Checks if input args are missing}
\usage{
fail_if_input_args_are_missing(args_list)
}
\arguments{
\item{args_list}{A named list.}
}
\value{
returns \code{args_list} invisibly.
}
\description{
Checks if any list entries in \code{args_list} are symbols. Called with
\code{args_list} as input argument this serves as a proxy for checking is
arguments are missing.
}
