% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_and_check.R
\name{fill_na_terms}
\alias{fill_na_terms}
\title{Fill missing terms with fallback_term}
\usage{
fill_na_terms(data, fallback_term)
}
\arguments{
\item{data}{A tibble holding at least column \code{term}.}

\item{fallback_term}{Numeric, holding fallback term.}
}
\value{
Tibble \code{data }
}
\description{
Also throws informative message on number of filled values.
}
