% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tech_trajectories.R
\name{filter_negative_late_and_sudden}
\alias{filter_negative_late_and_sudden}
\title{Remove negative late and sudden rows}
\usage{
filter_negative_late_and_sudden(data_with_late_and_sudden, log_path)
}
\arguments{
\item{data_with_late_and_sudden}{A tibble containing scenario data with
projected late and sudden trajectory.}

\item{log_path}{String holding path to log file.}
}
\value{
Input tibble with potentially removed rows.
}
\description{
Function checks for negative values on variable late_and_sudden. All
technology x company_name combinations holding >= 1 negative value are
removed.
}
