% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_transition_scenarios.R
\name{generate_litigation_shocks}
\alias{generate_litigation_shocks}
\title{Generate litigation scenario shock from a start year that represents when a
litigation settlement is reached.}
\usage{
generate_litigation_shocks(
  start_of_analysis,
  end_of_analysis,
  shock_year,
  scc,
  exp_share_damages_paid
)
}
\arguments{
\item{start_of_analysis}{A numeric vector of length one that indicates the
start year of the analysis.}

\item{end_of_analysis}{A numeric vector of length one that indicates the
end year of the analysis.}

\item{shock_year}{A numeric vector of length 1 that provides the start year
of the shock to be used in the analysis.}

\item{scc}{A numeric vector of length 1 that provides the social cost of
carbon}

\item{exp_share_damages_paid}{A numeric vector of length 1 that provides the
share of the liability expected to be paid out in a settlement}
}
\description{
Generate litigation scenario shock from a start year that represents when a
litigation settlement is reached.
}
