% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_price_data.R
\name{generate_price_data}
\alias{generate_price_data}
\title{Generates a toy data set of price data (long) for all relevant technologies}
\usage{
generate_price_data()
}
\value{
data frame
}
\description{
Does not take any inputs. Generates a data frame that contains toy price
data in the format generally expected in the climate transition risk stress
test. This is done for all technologies used in the stylized examples.
}
\seealso{
Other example case functions: 
\code{\link{generate_price_data_wide}()},
\code{\link{generate_test_capacity_factors}()},
\code{\link{generate_test_companies}()},
\code{\link{generate_test_exclude_companies_data}()},
\code{\link{generate_test_exposure}()},
\code{\link{generate_test_financial_data}()},
\code{\link{generate_test_scenarios}()}
}
\concept{example case functions}
