% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_capacity_factors.R
\name{generate_test_capacity_factors}
\alias{generate_test_capacity_factors}
\title{Generates a toy data set of capacity factors for the power technologies}
\usage{
generate_test_capacity_factors()
}
\value{
data frame
}
\description{
Does not take any inputs. Generates a data frame that contains toy capacity
factors data in the format generally expected in the climate transition risk
stress test. This is done for all technologies in the Power sector and the
scenarios used in the stylized examples.
}
\seealso{
Other example case functions: 
\code{\link{generate_price_data_wide}()},
\code{\link{generate_price_data}()},
\code{\link{generate_test_companies}()},
\code{\link{generate_test_exposure}()},
\code{\link{generate_test_scenarios}()}
}
\concept{example case functions}
