% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_companies.R
\name{generate_test_companies}
\alias{generate_test_companies}
\title{Generates a toy PACTA style portfolio with example companies}
\usage{
generate_test_companies(data)
}
\arguments{
\item{data}{data frame containing the test scenario data set}
}
\value{
data frame
}
\description{
Takes example scenario data set as input. Generates a PACTA style portfolio
that contains companies representing stylized cases of interest for the
transition risk stress test.
}
\seealso{
Other example case functions: 
\code{\link{generate_price_data_wide}()},
\code{\link{generate_price_data}()},
\code{\link{generate_test_capacity_factors}()},
\code{\link{generate_test_exclude_companies_data}()},
\code{\link{generate_test_exposure}()},
\code{\link{generate_test_financial_data}()},
\code{\link{generate_test_scenarios}()}
}
\concept{example case functions}
