% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_exclude_companies_data.R
\name{generate_test_exclude_companies_data}
\alias{generate_test_exclude_companies_data}
\title{Generates a toy empty exclude-companies data set.}
\usage{
generate_test_exclude_companies_data()
}
\value{
data frame
}
\description{
Does not take any inputs. Generates an data set that corresponds in structure
to the one used in the transition risk stress test. Company exclusion is not
a main test case, hence no entries are required.
}
\seealso{
Other example case functions: 
\code{\link{generate_price_data_wide}()},
\code{\link{generate_price_data}()},
\code{\link{generate_test_capacity_factors}()},
\code{\link{generate_test_companies}()},
\code{\link{generate_test_exposure}()},
\code{\link{generate_test_financial_data}()},
\code{\link{generate_test_scenarios}()}
}
\concept{example case functions}
