% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_exposure.R
\name{generate_test_exposure}
\alias{generate_test_exposure}
\title{Generates a toy PACTA style portfolio with example companies}
\usage{
generate_test_exposure(data)
}
\arguments{
\item{data}{data frame containing the test company/portfolio data set}
}
\value{
data frame
}
\description{
Takes example company/portfolio data set as input. Generates a PACTA style
exposure data frame. Contains the sector level financial exposures of the toy
portfolio to the companies representing stylized cases of interest for the
transition risk stress test.
}
\seealso{
Other example case functions: 
\code{\link{generate_price_data_wide}()},
\code{\link{generate_price_data}()},
\code{\link{generate_test_capacity_factors}()},
\code{\link{generate_test_companies}()},
\code{\link{generate_test_financial_data}()},
\code{\link{generate_test_scenarios}()}
}
\concept{example case functions}
