% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_financial_data.R
\name{generate_test_financial_data}
\alias{generate_test_financial_data}
\title{Generates a toy data set of financial information for example companies}
\usage{
generate_test_financial_data(data)
}
\arguments{
\item{data}{data frame containing the test financial data set}
}
\value{
data frame
}
\description{
Takes example company/portfolio data set as input. Generates a data frame
with toy financial data information. Contains relevant financial data for
profit calculation and Merton model for all companies for which test cases
are defined. Currently the financial data is identical for all companies, as
the test cases do not test variations in the credit risk part of the model.
}
\seealso{
Other example case functions: 
\code{\link{generate_price_data_wide}()},
\code{\link{generate_price_data}()},
\code{\link{generate_test_capacity_factors}()},
\code{\link{generate_test_companies}()},
\code{\link{generate_test_exposure}()},
\code{\link{generate_test_scenarios}()}
}
\concept{example case functions}
