% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_transition_scenarios.R
\name{generate_transition_shocks}
\alias{generate_transition_shocks}
\title{Generate transition scenario shock from a start year that represents when a
large scale climate transition policy is deployed.}
\usage{
generate_transition_shocks(start_of_analysis, end_of_analysis, shock_year)
}
\arguments{
\item{start_of_analysis}{A numeric vector of length one that indicates the
start year of the analysis.}

\item{end_of_analysis}{A numeric vector of length one that indicates the
end year of the analysis.}

\item{shock_year}{A numeric vector of length 1 that provides the start year
of the shock to be used in the analysis.}
}
\description{
Generate transition scenario shock from a start year that represents when a
large scale climate transition policy is deployed.
}
