% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-helpers.R
\name{geographies_for_sector}
\alias{geographies_for_sector}
\title{Obtain available geographies for sector}
\usage{
geographies_for_sector(sector, overview = scenario_geography_x_ald_sector)
}
\arguments{
\item{sector}{String of length 1 holding sector name.}

\item{overview}{A tibble holding valid combinations of \code{scenario_geography}
and \code{ald_sector}.}
}
\value{
A string vector holding supported scenario_geographies.
}
\description{
Function returns a vector holding names of scenario_geographies for a
provided \code{sector}. Source of truth is the \code{overview}
\code{scenario_geography_x_ald_sector} per default. In case \code{sector} is not
available an error is thrown.
}
\examples{
geographies_for_sector("Coal")
}
