% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_iter_var}
\alias{get_iter_var}
\title{Get name of iterator variable}
\usage{
get_iter_var(args_list)
}
\arguments{
\item{args_list}{Named list of default and provided arguments in function
call to \code{\link[=run_trisk]{run_trisk()}}.}
}
\value{
String holding name of iterator variable.
}
\description{
Uses fallback if no iterator is used. Aborts if > 1 iterator is given.
}
