% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-helpers.R
\name{get_scenario_geography_x_ald_sector}
\alias{get_scenario_geography_x_ald_sector}
\title{Get scenario_geography_x_ald_sector}
\usage{
get_scenario_geography_x_ald_sector(st_input_folder, whitelist_sectors = NULL)
}
\arguments{
\item{st_input_folder}{path to the folder containing the stress test files}

\item{whitelist_sectors}{only sectors to use if any (leave NULL to use all sectors)}
}
\value{
scenario_geography_x_ald_sector
}
\description{
Get scenario_geography_x_ald_sector
}
