% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inner_join_report_drops}
\alias{inner_join_report_drops}
\title{Inner join datasets and report number of dropped rows}
\usage{
inner_join_report_drops(data_x, data_y, name_x, name_y, merge_cols)
}
\arguments{
\item{data_x}{Tibble with data that is joinable to \code{data_y}.}

\item{data_y}{Tibble with data that is joinable to \code{data_x}.}

\item{name_x}{Name of \code{data_x}.}

\item{name_y}{Name of \code{data_x}.}

\item{merge_cols}{Vector holds columns to join on.}
}
\value{
The merged dataset.
}
\description{
Function conducts inner join on two datasets and reports number of dropped
rows.
}
