% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tests.R
\name{is_registered_dev}
\alias{is_registered_dev}
\title{Allow running tests selectively, only for registered developers}
\usage{
is_registered_dev(my_home = "/home/mauro")
}
\value{
Logical.
}
\description{
Allow running tests selectively, only for registered developers
}
\examples{
is_registered_dev()
is_registered_dev("home/you/")
}
