% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\docType{data}
\name{lookup}
\alias{lookup}
\alias{nesting_vars_lookup}
\alias{asset_types_lookup}
\alias{scenarios_lookup}
\alias{baseline_scenario_lookup}
\alias{shock_scenario_lookup}
\alias{calculation_level_lookup}
\alias{lgd_senior_claims_range_lookup}
\alias{lgd_subordinated_claims_range_lookup}
\alias{terminal_value_range_lookup}
\alias{risk_free_rate_range_lookup}
\alias{discount_rate_range_lookup}
\alias{div_netprofit_prop_coef_range_lookup}
\alias{shock_year_range_lookup}
\alias{term_range_lookup}
\alias{end_year_lookup}
\alias{credit_type_lookup}
\alias{sectors_lookup}
\alias{technologies_lookup}
\alias{allocation_method_lookup}
\alias{p4i_p4b_sector_technology_lookup}
\alias{p4i_p4b_scenario_lookup}
\alias{p4b_scenarios_lookup}
\alias{p4b_production_data_years_lookup}
\alias{p4b_scenario_data_years_lookup}
\title{Help lookup values}
\format{
An object of class \code{character} of length 9.

An object of class \code{character} of length 3.

An object of class \code{character} of length 4.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{numeric} of length 2.

An object of class \code{numeric} of length 2.

An object of class \code{numeric} of length 2.

An object of class \code{numeric} of length 2.

An object of class \code{numeric} of length 2.

An object of class \code{numeric} of length 2.

An object of class \code{numeric} of length 2.

An object of class \code{numeric} of length 2.

An object of class \code{numeric} of length 1.

An object of class \code{character} of length 2.

An object of class \code{character} of length 4.

An object of class \code{character} of length 12.

An object of class \code{character} of length 1.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 2 columns.

An object of class \code{character} of length 7.

An object of class \code{numeric} of length 2.

An object of class \code{numeric} of length 2.
}
\usage{
nesting_vars_lookup

asset_types_lookup

scenarios_lookup

baseline_scenario_lookup

shock_scenario_lookup

calculation_level_lookup

lgd_senior_claims_range_lookup

lgd_subordinated_claims_range_lookup

terminal_value_range_lookup

risk_free_rate_range_lookup

discount_rate_range_lookup

div_netprofit_prop_coef_range_lookup

shock_year_range_lookup

term_range_lookup

end_year_lookup

credit_type_lookup

sectors_lookup

technologies_lookup

allocation_method_lookup

p4i_p4b_sector_technology_lookup

p4i_p4b_scenario_lookup

p4b_scenarios_lookup

p4b_production_data_years_lookup

p4b_scenario_data_years_lookup
}
\description{
Help lookup values
}
\examples{
nesting_vars_lookup

asset_types_lookup

scenarios_lookup

baseline_scenario_lookup

shock_scenario_lookup

calculation_level_lookup

lgd_senior_claims_range_lookup

lgd_subordinated_claims_range_lookup

terminal_value_range_lookup

risk_free_rate_range_lookup

discount_rate_range_lookup

div_netprofit_prop_coef_range_lookup

shock_year_range_lookup

term_range_lookup

end_year_lookup

credit_type_lookup

sectors_lookup

technologies_lookup

allocation_method_lookup

p4i_p4b_sector_technology_lookup

p4i_p4b_scenario_lookup

p4b_scenarios_lookup

p4b_production_data_years_lookup

p4b_scenario_data_years_lookup
}
\keyword{datasets}
