% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall_pd_change_company_technology.R
\name{overall_pd_change_company_technology}
\alias{overall_pd_change_company_technology}
\title{Display the overall pd changes per maturity, company and technology.}
\usage{
overall_pd_change_company_technology(
  data,
  shock_year = NULL,
  sector_filter = NULL,
  company_filter = NULL,
  geography_filter = NULL
)
}
\arguments{
\item{data}{A data.frame that contains Merton output data.}

\item{shock_year}{Numeric. Vector of length 1 that specifies the year of the
policy shock.}

\item{sector_filter}{Character. A vector of sectors to display.}

\item{company_filter}{Character. A vector of companies to display.}

\item{geography_filter}{Character. A vector containing the geographical
specification of the scenario}
}
\value{
ggplot object
}
\description{
Display the overall pd changes per maturity, company and technology.
}
\seealso{
Other qa graphing functions: 
\code{\link{annual_pd_change_company_technology}()},
\code{\link{annual_pd_change_technology_shock_year}()},
\code{\link{overall_pd_change_technology_shock_year}()},
\code{\link{show_carbon_budget}()},
\code{\link{show_impact_by_shock_year}()},
\code{\link{show_pf_technology_shares}()},
\code{\link{show_price_trajectories}()},
\code{\link{show_prod_baseline_target_ls_pf}()},
\code{\link{show_prod_trajectories}()},
\code{\link{show_var_change_by_shock_year}()}
}
\concept{qa graphing functions}
