% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paste_write}
\alias{paste_write}
\title{Helper function for logging}
\usage{
paste_write(..., log_path)
}
\arguments{
\item{...}{One or more R objects, to be converted to character vectors.}

\item{log_path}{String holding path to logfile.}
}
\description{
Concatenates onjects passed in \code{...} and writes resulting string to
\code{logfile}.
}
