% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path_dropbox_2dii}
\alias{path_dropbox_2dii}
\title{Easily access directories in your local copy of 2dii's Dropbox folder}
\usage{
path_dropbox_2dii(...)
}
\arguments{
\item{...}{Character vectors, if any values are \code{NA}, the result will also be
\code{NA}.}
}
\value{
A character string.
}
\description{
These functions create cross-platform paths pointing to 2dii's Dropbox
folder:
Your projects may need data stored in 2dii's Dropbox folder. Sometimes it is
convenient to have your projects close to the data. But, in this case, it is
a bad idea because the path to 2dii's Dropbox folder has a problematic space
and symbol. (For example, \href{https://github.com/2DegreesInvesting/resources/issues/51}{RStudio's Git pane may not work}.) Instead,
place your projects somewhere with a sane path, such as
\verb{C:/Users/You/git/project/}, and access the data you need with
\code{path_dropbox_2dii()}.
}
\section{Setup for a custom Dropbox folder}{

If the name of your 2dii Dropbox folder is different from the default,
you may add this to .Rprofile (see \code{\link[usethis:edit]{usethis::edit_r_profile()}}):\preformatted{options(r2dii_dropbox = "The name of your custom dropbox folder goes here")
}
}

\seealso{
\code{\link[=degrees]{degrees()}}, \code{\link[fs:path_expand]{fs::path_home()}}.
}
\concept{functions to output 2dii paths}
