% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_company_terms}
\alias{process_company_terms}
\title{Process data of type indicated by function name}
\usage{
process_company_terms(data, fallback_term)
}
\arguments{
\item{data}{A tibble of data of type indicated by function name.}

\item{fallback_term}{Numeric, holding fallback term.}
}
\value{
A tibble of data as indicated by function name.
}
\description{
NOTE returns NULL if \code{data} is NULL.
}
