% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_pacta_results}
\alias{process_pacta_results}
\title{Process data of type indicated by function name}
\usage{
process_pacta_results(
  data,
  start_year,
  end_year,
  time_horizon,
  scenario_geography_filter,
  scenarios_filter,
  equity_market_filter,
  sectors,
  technologies,
  allocation_method,
  asset_type
)
}
\arguments{
\item{data}{A tibble of data of type indicated by function name.}

\item{start_year}{Numeric, holding start year of analysis.}

\item{end_year}{Numeric, holding end year of analysis.}

\item{time_horizon}{Numeric, holding time horizon of analysis.}

\item{scenario_geography_filter}{Character. A vector of length 1 that
indicates which geographic scenario to apply in the analysis.}

\item{scenarios_filter}{Vector holding baseline and shock scenario name.}

\item{equity_market_filter}{Character. A vector of length 1 that
indicates which equity market to apply in the analysis.}

\item{sectors}{Character vector, holding considered sectors.}

\item{technologies}{Character vector, holding considered technologies.}

\item{allocation_method}{Character. A vector of length 1 indicating the
set of PACTA data to be used in the analysis, based on the choice of an
allocation rule.}

\item{asset_type}{String holding asset_type. For accepted values compare
\code{stress_test_arguments}.}
}
\value{
A tibble of data as indicated by function name.
}
\description{
Process data of type indicated by function name
}
