% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_production_data}
\alias{process_production_data}
\title{Process data of type indicated by function name}
\usage{
process_production_data(
  data,
  start_year,
  end_year,
  time_horizon,
  scenario_geography_filter,
  sectors,
  technologies,
  log_path
)
}
\arguments{
\item{data}{A tibble of data of type indicated by function name.}

\item{start_year}{Numeric, holding start year of analysis.}

\item{end_year}{Numeric, holding end year of analysis.}

\item{time_horizon}{Numeric, holding time horizon of production data.}

\item{scenario_geography_filter}{Character. A vector of length 1 that
indicates which geographic scenario to apply in the analysis.}

\item{sectors}{Character vector, holding considered sectors.}

\item{technologies}{Character vector, holding considered technologies.}

\item{log_path}{String holding path to log file.}
}
\value{
A tibble of data as indicated by function name.
}
\description{
Process data of type indicated by function name
}
