% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_prepare.R
\name{read_and_prepare_project_agnostic_data}
\alias{read_and_prepare_project_agnostic_data}
\title{Read and prepare project agnostic data}
\usage{
read_and_prepare_project_agnostic_data(
  start_year,
  end_year,
  company_exclusion,
  scenario_geography_filter,
  asset_type,
  path
)
}
\arguments{
\item{start_year}{Start year of analysis.}

\item{end_year}{End year of analysis.}

\item{company_exclusion}{Boolean, indicating if companies provided in dataset
excluded_companies.csv shall be excluded.}

\item{scenario_geography_filter}{String holding geographic asset locations to
be considered in analysis.}

\item{asset_type}{Type of asset, can be bonds, loans or equity.}

\item{path}{String holding path to data.}
}
\value{
A list holding prepared project agnostic data.
}
\description{
Function reads in data that are independent of the project and conducts some
checking and wrangling.
}
