% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_capacity_factors.R
\name{read_capacity_factors}
\alias{read_capacity_factors}
\title{Read in power capacity factors from csv and check that all expected columns
are given.}
\usage{
read_capacity_factors(path = NULL)
}
\arguments{
\item{path}{A string that points to the location of the file containing the
capacity factors}
}
\description{
Read in power capacity factors from csv and check that all expected columns
are given.
}
\concept{import functions}
