% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_carbon_data.R
\name{read_carbon_data}
\alias{read_carbon_data}
\title{Read in carbon price data from ngfs data}
\usage{
read_carbon_data()
}
\arguments{
\item{path}{A string that points to the location of the file containing the
carbon price data.}
}
\description{
Read in carbon price data from ngfs data
}
\seealso{
Other import functions: 
\code{\link{read_capacity_factors_power}()},
\code{\link{read_financial_data}()},
\code{\link{read_production_data}()}
}
\concept{import functions}
