% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_company_data.R
\name{read_company_data}
\alias{read_company_data}
\title{Read in company finacial data processed from eikon exports and AR master data
that contain information on multiple credit risk inputs and company
production plans.}
\usage{
read_company_data(path = NULL)
}
\arguments{
\item{path}{A string that points to the location of the file containing the
company financial data.}
}
\description{
Read in company finacial data processed from eikon exports and AR master data
that contain information on multiple credit risk inputs and company
production plans.
}
\seealso{
Other import functions: 
\code{\link{read_capacity_factors}()},
\code{\link{read_ngfs_carbon_tax}()},
\code{\link{read_pacta_results}()},
\code{\link{read_transition_scenarios}()}
}
\concept{import functions}
