% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excluded_companies.R
\name{read_excluded_companies}
\alias{read_excluded_companies}
\title{Read excluded company data}
\usage{
read_excluded_companies(path)
}
\arguments{
\item{path}{Path to company exclusion data.}
}
\value{
A tibble holding company exclusion data.
}
\description{
Read excluded company data
}
