% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_price_data.R
\name{read_price_data_internal}
\alias{read_price_data_internal}
\title{Read in price data from csv and check that all expected columns are given.}
\usage{
read_price_data_internal(path)
}
\arguments{
\item{path}{A string that points to the location of the file containing the
price data}
}
\description{
This function reads in price data using long file format.
It is expected to work with data based on IEA WEO 2020.
}
