% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_price_data_internal_old.R
\name{read_price_data_internal_old}
\alias{read_price_data_internal_old}
\title{Read in price data from csv and check that all expected columns are given.}
\usage{
read_price_data_internal_old(path)
}
\arguments{
\item{path}{A string that points to the location of the file containing the
price data}
}
\description{
This function reads in price data using the old wide data format.
}
