% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_price_data.R
\name{read_price_data_old}
\alias{read_price_data_old}
\title{Read in price data}
\usage{
read_price_data_old(path, expected_technologies)
}
\arguments{
\item{path}{A string that points to the location of the file containing the
price data}

\item{expected_technologies}{String vector holding expected technologies.}
}
\value{
A tibble holding price data in long format.
}
\description{
This function reads in price data using the old wide data format.
}
