% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_price_data.R
\name{read_price_data_old2}
\alias{read_price_data_old2}
\title{Read in price data}
\usage{
read_price_data_old2(path)
}
\arguments{
\item{path}{A string that points to the location of the file containing the
price data}
}
\value{
A tibble holding price data in long format.
}
\description{
This function reads in price data using the old wide data format.
}
