% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_production_data.R
\name{read_production_data}
\alias{read_production_data}
\title{Read in AR PAMS production data.}
\usage{
read_production_data(path = NULL)
}
\arguments{
\item{path}{A string that points to the location of the file containing the
AR PAMS dataset}
}
\description{
Read in AR PAMS production data.
}
\seealso{
Other import functions: 
\code{\link{read_capacity_factors_power}()},
\code{\link{read_financial_data}()}
}
\concept{import functions}
