% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_scenario_data.R
\name{read_scenario_data}
\alias{read_scenario_data}
\title{Read in scenario data}
\usage{
read_scenario_data(path)
}
\arguments{
\item{path}{Path to dir holding scenario data.}
}
\value{
A tibble holding scenario data.
}
\description{
Function reads in scenario data and checks for existence of required columns.
}
