% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_prepare.R
\name{read_sector_exposures}
\alias{read_sector_exposures}
\title{Read sector exposures}
\usage{
read_sector_exposures(path)
}
\arguments{
\item{path}{Path to file holding sector exposures.}
}
\value{
A tibble holding sector exposure data
}
\description{
Read file holding sector exposure results.
}
