% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_transition_scenarios.R
\name{read_transition_scenarios}
\alias{read_transition_scenarios}
\title{Read in transition scenarios from csv and check that all expected columns
are given.}
\usage{
read_transition_scenarios(
  path = NULL,
  start_of_analysis = NULL,
  end_of_analysis = NULL
)
}
\arguments{
\item{path}{A string that points to the location of the file containing the
transition scenarios}

\item{start_of_analysis}{A numeric vector of length one that indicates the
start year of the analysis}

\item{end_of_analysis}{A numeric vector of length one that indicates the
end year of the analysis}
}
\description{
Read in transition scenarios from csv and check that all expected columns
are given.
}
\seealso{
Other import functions: 
\code{\link{generate_transition_shocks}()},
\code{\link{read_capacity_factors_old}()},
\code{\link{read_capacity_factors}()},
\code{\link{read_financial_data}()},
\code{\link{read_ngfs_carbon_tax}()},
\code{\link{read_pacta_results}()},
\code{\link{read_price_data}()}
}
\concept{import functions}
