% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_and_check.R
\name{rename_results}
\alias{rename_results}
\title{Rename results}
\usage{
rename_results(results_list)
}
\arguments{
\item{results_list}{A list of wrangled and checked results.}
}
\value{
\code{result_list} with adjusted named
}
\description{
Rename results lists entries so that results as returned to the user on
demand follow naming convention of exported results.
}
