% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{report_all_duplicate_kinds}
\alias{report_all_duplicate_kinds}
\title{Checks data for missings and duplicates}
\usage{
report_all_duplicate_kinds(data, composite_unique_cols)
}
\arguments{
\item{data}{A tibble.}

\item{composite_unique_cols}{A vector of names of columns that shall be
unique in their combination.}
}
\description{
Applies consistency checks to data concerning the combinations of columns
that should be unique in combination. In concrete:
}
\details{
\enumerate{
\item it is checked if there are duplicate rows.
\item it is checked if there are duplicate rows on \code{composite_unique_cols}.
}

Function is currently not used in code but helps for data wrangling/data
research tasks. It will be added to critical data sets in the future.
}
