% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{report_and_remove_duplicates}
\alias{report_and_remove_duplicates}
\title{Report and remove duplicate rows}
\usage{
report_and_remove_duplicates(data, cols)
}
\arguments{
\item{data}{A tibble.}

\item{cols}{Cols to check for duplicate combinations on}
}
\value{
Tibble \code{data} with removed duplicates.
}
\description{
Reports and removes duplicates in \code{data} on columns \code{cols}.
}
