% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{report_company_drops}
\alias{report_company_drops}
\title{Report dropped companies}
\usage{
report_company_drops(data_list, asset_type, log_path)
}
\arguments{
\item{data_list}{A list of imported stress test input data.}

\item{asset_type}{String holding asset_type. For accepted values compare
\code{stress_test_arguments}.}

\item{log_path}{String holding path to log file.}
}
\description{
Wrapper to report companies for which all results, or results for some technologies
are lsot due to a missing match in financial_data or price_data.
}
