% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{report_dropped_company_names}
\alias{report_dropped_company_names}
\title{Inner join datasets and report number of dropped rows}
\usage{
report_dropped_company_names(
  data_x,
  data_y,
  name_y,
  merge_cols,
  name_x = "PACTA results",
  log_path
)
}
\arguments{
\item{data_x}{Tibble with data that is joinable to \code{data_y}.}

\item{data_y}{Tibble with data that is joinable to \code{data_x}.}

\item{name_y}{Name of \code{data_x}.}

\item{merge_cols}{Vector holds columns to join on.}

\item{name_x}{Name of `data_x, defults to PACTA results.}

\item{log_path}{String holding path to log file.}
}
\value{
The merged dataset.
}
\description{
Function conducts inner join on two datasets and reports number of dropped
rows on \code{data_x}.
}
