% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{report_missings}
\alias{report_missings}
\title{Report missing}
\usage{
report_missings(data, name_data)
}
\arguments{
\item{data}{Tibble holding a result data set.}

\item{name_data}{Name of the data file.}
}
\description{
Function reports number of missing values per variable.
}
