% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{report_missings_col_combinations}
\alias{report_missings_col_combinations}
\title{Identify and report missing value combinations}
\usage{
report_missings_col_combinations(data, composite_unique_cols)
}
\arguments{
\item{data}{A tibble.}

\item{composite_unique_cols}{A vector of names of columns that shall be
unique in their combination.}
}
\description{
Identifies and reports missing value combinations in \code{data} on
\code{composite_unique_cols}.
}
