% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun_st.R
\name{rerun_st}
\alias{rerun_st}
\title{Sensitivity analysis: \code{run_stress_test()} multiple times}
\usage{
rerun_st(data, asset_type, ..., quiet = TRUE)
}
\arguments{
\item{data}{Character vector of length 2. Paths to the directories
ST_INPUTS_MASTER and ST_TESTING_\{aset-type\}, e.g. ST_TESTING_BONDS,
respectively. These directories must have the files expected by the
stress-test project.}

\item{asset_type}{An atomic character vector. Either "bonds", "equity", or
"loans".}

\item{...}{Arguments passed to \code{\link[=run_stress_test]{run_stress_test()}}.}

\item{quiet}{Logical. Print non-condiiton messages to the console?}
}
\value{
A data frame.
}
\description{
This function performs the so called "sensitivity analysis". It runs
\code{\link[=run_stress_test]{run_stress_test()}} multiple times, iterating over multiple values of one and
only one argument of \code{\link[=run_stress_test]{run_stress_test()}}.
}
\examples{
\dontshow{if (r2dii.climate.stress.test:::is_registered_dev()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(dplyr, warn.conflicts = FALSE)
library(readr, warn.conflicts = FALSE)
library(tidyr, warn.conflicts = FALSE)

data <- st_data_paths(
  data = "/home/mauro/tmp/st/ST_INPUTS_MASTER",
  project = "/home/mauro/tmp/st/ST_TESTING_BONDS"
)

compact <- suppressWarnings(
  rerun_st(data, "bonds", term = c(1, 2, 3))
)

# A compact view of all your results
compact

# The full view of all your results
full <- unnest(compact, st_result)

# You may save the result
path <- tempfile()
write_csv(full, file = path)

# And eventually re-read it
read_csv(path, show_col_types = FALSE)

# Or explore interesting results
filter(full, st_name == "port", arg_value == 2)
\dontshow{\}) # examplesIf}
}
\seealso{
Other main functions: 
\code{\link{run_stress_test}()}
}
\concept{main functions}
