% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stress_test-addin.R
\name{run_stress_test_addin}
\alias{run_stress_test_addin}
\title{Run \code{run_stress_test()}}
\usage{
run_stress_test_addin()
}
\description{
\code{run_stress_test_addin()} opens an \href{https://shiny.rstudio.com/articles/gadgets.html}{RStudio gadget} and
\href{http://rstudio.github.io/rstudioaddins/}{addin} that allows you to set the
paths to inputs and outputs. Appears as "Run stress test" in the RStudio
Addins menu.
}
