% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stress_test.R
\name{run_stress_test_impl}
\alias{run_stress_test_impl}
\title{Run stress testing for provided asset type.}
\usage{
run_stress_test_impl(
  asset_type,
  input_path_project_specific,
  input_path_project_agnostic,
  output_path,
  lgd_senior_claims,
  lgd_subordinated_claims,
  terminal_value,
  risk_free_rate,
  discount_rate,
  div_netprofit_prop_coef,
  shock_year,
  term,
  company_exclusion,
  iter_var
)
}
\arguments{
\item{asset_type}{String holding asset_type, for allowed value compare
\code{asset_types_lookup}.}

\item{input_path_project_specific}{String holding path to project specific
data.}

\item{input_path_project_agnostic}{String holding path to project agnostic
data.}

\item{output_path}{String holding path to which output files are written.}

\item{lgd_senior_claims}{Numeric, holding the loss given default for senior
claims, for accepted value range check \code{lgd_senior_claims_range_lookup}.}

\item{lgd_subordinated_claims}{Numeric, holding the loss given default for
subordinated claims, for accepted value range check
\code{lgd_subordinated_claims_range_lookup}.}

\item{terminal_value}{Numeric. A ratio to determine the share of the
discounted value used in the terminal value calculation beyond the
projected time frame. For accepted range compare
\code{terminal_value_range_lookup}.}

\item{risk_free_rate}{Numeric that indicates the risk free rate of interest.
For accepted range compare \code{risk_free_rate_range_lookup}.}

\item{discount_rate}{Numeric, that holds the discount rate of dividends per
year in the DCF. For accepted range compare \code{discount_rate_range_lookup}.}

\item{div_netprofit_prop_coef}{Numeric. A coefficient that determines how
strongly the future dividends propagate to the company value. For accepted
range compare \code{div_netprofit_prop_coef_range_lookup}.}

\item{shock_year}{Numeric, holding year the shock is applied. For accepted
range compare \code{shock_year_range_lookup}.}

\item{term}{Numeric. A coefficient that determines for which maturity the
expected loss should be calculated in the credit risk section. For accepted
range compare \code{term_range_lookup}.}

\item{company_exclusion}{Boolean, indicating if companies provided in dataset
excluded_companies.csv shall be excluded.}

\item{iter_var}{String holding name of iteration variable.}
}
\value{
A list of stress test results.
}
\description{
Runs stress test per iteration.
}
