% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stress_test.R
\name{run_stress_test_iteration}
\alias{run_stress_test_iteration}
\title{Iterate over stress test runs}
\usage{
run_stress_test_iteration(n, args_tibble)
}
\arguments{
\item{n}{Numeric.}

\item{args_tibble}{A tibble holding a set of params for
\code{run_stress_test_imp} per row.}
}
\value{
List of stress test results.
}
\description{
Iterate over stress test runs
}
