% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_stress_test_loans.R
\name{run_stress_test_loans}
\alias{run_stress_test_loans}
\title{Run stress testing for loans}
\usage{
run_stress_test_loans(
  lgd_senior_claims = 0.45,
  lgd_subordinated_claims = 0.75,
  terminal_value = 0,
  risk_free_rate = 0.02,
  discount_rate = 0.02,
  div_netprofit_prop_coef = 1,
  shock_year = 2030,
  term = 2,
  company_exclusion = TRUE,
  credit_type = "outstanding"
)
}
\arguments{
\item{lgd_senior_claims}{Numeric, holding the loss given default for senior
claims, for accepted value range check \code{lgd_senior_claims_range_lookup}.}

\item{lgd_subordinated_claims}{Numeric, holding the loss given default for
subordinated claims, for accepted value range check
\code{lgd_subordinated_claims_range_lookup}.}

\item{terminal_value}{Numeric. A ratio to determine the share of the
discounted value used in the terminal value calculation beyond the
projected time frame. For accepted range compare \code{terminal_value_range_lookup}.}

\item{risk_free_rate}{Numeric that indicates the risk free rate of interest.
For accepted range compare \code{risk_free_rate_range_lookup}.}

\item{discount_rate}{Numeric, that holds the discount rate of dividends per
year in the DCF. For accepted range compare \code{discount_rate_range_lookup}.}

\item{div_netprofit_prop_coef}{Numeric. A coefficient that determines how
strongly the future dividends propagate to the company value. For accepted
range compare \code{div_netprofit_prop_coef_range_lookup}.}

\item{shock_year}{Numeric, holding year the shock is applied. For accepted
range compare \code{shock_year_range_lookup}.}

\item{term}{Numeric. A coefficient that determines for which maturity the
expected loss should be calculated in the credit risk section. For accepted
range compare \code{term_range_lookup}.}

\item{company_exclusion}{Boolean, indicating if companies provided in dataset
excluded_companies.csv shall be excluded.}

\item{credit_type}{Type of credit. For accepted values please compare
\code{credit_type_loans}.}
}
\description{
Run stress testing for loans
}
