% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-helpers.R
\name{scenario_for_sector_x_geography}
\alias{scenario_for_sector_x_geography}
\title{Obtain available scenario_x_source for geography - sector combinations}
\usage{
scenario_for_sector_x_geography(st_input_folder, sector, scenario_geography)
}
\arguments{
\item{st_input_folder}{path to the input folder containing stress test input files}

\item{sector}{String of length 1 holding sector name.}

\item{scenario_geography}{String of length 1 holding sector name.}
}
\value{
A string holding valid scenario names.
}
\description{
Function returns a vector holding names of scenarios for a provided \code{sector}
and \code{scenario_geography}. To identify valid \code{sector} x \code{scenario_geography}
combinations refer to \code{\link[=geographies_for_sector]{geographies_for_sector()}}. Source of truth is the
\code{overview} \code{scenario_geography_x_ald_sector} per default. In case combination
of \code{sector} and \code{scenario_geography} is not available an error is thrown.
}
