% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_geography_x_ald_sector.R
\docType{data}
\name{scenario_geography_x_ald_sector}
\alias{scenario_geography_x_ald_sector}
\title{A dataset describing valid combinations of scenario_geography and sector}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 101 rows and 3 columns.
}
\usage{
scenario_geography_x_ald_sector
}
\description{
A dataset describing valid combinations of scenario_geography and sector
}
\examples{
scenario_geography_x_ald_sector
}
\keyword{datasets}
