% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_and_check.R
\name{select_sector_scenario_combinations}
\alias{select_sector_scenario_combinations}
\title{Select required sector scenario combinations}
\usage{
select_sector_scenario_combinations(pacta_results)
}
\arguments{
\item{pacta_results}{A tibble holding pacta results}
}
\value{
Tibble \code{pacta_results} holding only required sector - scenario
combinations.
}
\description{
Function:
\enumerate{
\item Checks that for all provided sector required scenarios are available
\item Filters so that only required scenarios are kept.
}
}
