% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_carbon_budget.R
\name{show_carbon_budget}
\alias{show_carbon_budget}
\title{Display the carbon budget per technology over time}
\usage{
show_carbon_budget(
  data,
  scenarios = NULL,
  target_scenario = NULL,
  scenario_name_qa = "Carbon balance 2030",
  cumulative = FALSE
)
}
\arguments{
\item{data}{A dataframe that contains the QA annual profits data frame for an
asset type.}

\item{scenarios}{A dataframe that contains the fully prepared scenario data
used in the analysis.}

\item{target_scenario}{Character. A vector of length one containing the name
of the target scenario against which to compare the carbon budget.}

\item{scenario_name_qa}{Chracter. A vector of length one containing the name of
the transition scenario to inspect.}

\item{cumulative}{Logical. A vector of length one, indicating whether the
output should be an overview of changes in the carbon budget per technology
over the time frame of the analysis (FALSE) or a summary that shows the
overall deviation from the carbon budget by technology (TRUE).}
}
\value{
ggplot object
}
\description{
Display the carbon budget per technology over time
}
\seealso{
Other qa graphing functions: 
\code{\link{show_impact_by_shock_year}()},
\code{\link{show_pf_technology_shares}()},
\code{\link{show_price_trajectories}()},
\code{\link{show_prod_baseline_target_ls_pf}()},
\code{\link{show_prod_trajectories}()},
\code{\link{show_var_change_by_shock_year}()}
}
\concept{qa graphing functions}
