% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_graphs_st.R
\name{show_impact_by_shock_year}
\alias{show_impact_by_shock_year}
\title{Display impact on value at risk by technology and shock year}
\usage{
show_impact_by_shock_year(data, level = NULL)
}
\arguments{
\item{data}{A dataframe that contains stress test results for one asset type.}

\item{level}{Character. A string, either "technology" or "ald_sector",
indicating the desired level of aggregation.}
}
\value{
ggplot object
}
\description{
Display impact on value at risk by technology and shock year
}
\seealso{
Other qa graphing functions: 
\code{\link{show_carbon_budget}()},
\code{\link{show_pf_technology_shares}()},
\code{\link{show_price_trajectories}()},
\code{\link{show_prod_baseline_target_ls_pf}()},
\code{\link{show_prod_trajectories}()},
\code{\link{show_var_change_by_shock_year}()}
}
\concept{qa graphing functions}
