% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_graphs_st.R
\name{show_pf_technology_shares}
\alias{show_pf_technology_shares}
\title{Check the technology shares}
\usage{
show_pf_technology_shares(data)
}
\arguments{
\item{data}{A dataframe that contains the technology share of the portfolio
for one of the asset types.}
}
\value{
ggplot object
}
\description{
Check the technology shares
}
\seealso{
Other qa graphing functions: 
\code{\link{annual_pd_change_company_technology}()},
\code{\link{annual_pd_change_technology_shock_year}()},
\code{\link{overall_pd_change_company_technology}()},
\code{\link{overall_pd_change_technology_shock_year}()},
\code{\link{show_carbon_budget}()},
\code{\link{show_impact_by_shock_year}()},
\code{\link{show_price_trajectories}()},
\code{\link{show_prod_baseline_target_ls_pf}()},
\code{\link{show_prod_trajectories}()},
\code{\link{show_var_change_by_shock_year}()}
}
\concept{qa graphing functions}
