% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_graphs_st.R
\name{show_price_trajectories}
\alias{show_price_trajectories}
\title{Display the market price trajectories per technology}
\usage{
show_price_trajectories(
  data = df_price,
  price_scenarios = c("B2DS", "NPS", "SDS", "Baseline")
)
}
\arguments{
\item{data}{A dataframe that contains market prices for a list of scenarios.}

\item{price_scenarios}{Character. A vector containing the names of the
scenarios to display in the chart.}
}
\value{
ggplot object
}
\description{
Display the market price trajectories per technology
}
\seealso{
Other qa graphing functions: 
\code{\link{annual_pd_change_company_technology}()},
\code{\link{annual_pd_change_technology_shock_year}()},
\code{\link{overall_pd_change_company_technology}()},
\code{\link{overall_pd_change_technology_shock_year}()},
\code{\link{show_carbon_budget}()},
\code{\link{show_impact_by_shock_year}()},
\code{\link{show_pf_technology_shares}()},
\code{\link{show_prod_baseline_target_ls_pf}()},
\code{\link{show_prod_trajectories}()},
\code{\link{show_var_change_by_shock_year}()}
}
\concept{qa graphing functions}
