% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_graphs_st.R
\name{show_prod_baseline_target_ls_pf}
\alias{show_prod_baseline_target_ls_pf}
\title{Compare the production trajectories per technology by baseline, target and
late_and_sudden scenarios}
\usage{
show_prod_baseline_target_ls_pf(
  data,
  geography_filter = NULL,
  shock_year = NULL
)
}
\arguments{
\item{data}{A dataframe that contains the annual_profits output.}

\item{geography_filter}{Character. A vector of included scenario geographies.}

\item{shock_year}{Numeric. Indicates the year of shock for which to display
the production trajectories.}
}
\value{
ggplot object
}
\description{
Compare the production trajectories per technology by baseline, target and
late_and_sudden scenarios
}
\seealso{
Other qa graphing functions: 
\code{\link{show_carbon_budget}()},
\code{\link{show_impact_by_shock_year}()},
\code{\link{show_pf_technology_shares}()},
\code{\link{show_price_trajectories}()},
\code{\link{show_prod_trajectories}()},
\code{\link{show_var_change_by_shock_year}()}
}
\concept{qa graphing functions}
