% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_graphs_st.R
\name{show_prod_trajectories}
\alias{show_prod_trajectories}
\title{Display the market production trajectories per technology}
\usage{
show_prod_trajectories(
  data = scenario_data,
  end_year = 2040,
  source = NULL,
  ald_sector = NULL,
  technology = NULL,
  geography_filter = NULL
)
}
\arguments{
\item{data}{A dataframe that contains scenario data.}

\item{end_year}{Numeric. Final year of the analysis/trajectories.}

\item{source}{Character. A vector of allowed scenario sources.}

\item{ald_sector}{Character. A vector of allowed sectors.}

\item{technology}{Character. A vector of allowed technologies.}

\item{geography_filter}{Character. A vector of included scenario geographies.}
}
\value{
ggplot object
}
\description{
Display the market production trajectories per technology
}
\seealso{
Other qa graphing functions: 
\code{\link{annual_pd_change_company_technology}()},
\code{\link{annual_pd_change_technology_shock_year}()},
\code{\link{overall_pd_change_company_technology}()},
\code{\link{overall_pd_change_technology_shock_year}()},
\code{\link{show_carbon_budget}()},
\code{\link{show_impact_by_shock_year}()},
\code{\link{show_pf_technology_shares}()},
\code{\link{show_price_trajectories}()},
\code{\link{show_prod_baseline_target_ls_pf}()},
\code{\link{show_var_change_by_shock_year}()}
}
\concept{qa graphing functions}
