% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{source_all}
\alias{source_all}
\title{Source scripts quietly based on a list of input paths}
\usage{
source_all(paths)
}
\arguments{
\item{paths}{A vector of paths to scripts that need to be sourced for running
the work flow at hand. Usually this will point to files that contain
function definitions.}
}
\description{
Source scripts quietly based on a list of input paths
}
