% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_data_paths.R
\name{st_data_paths}
\alias{st_data_paths}
\title{Path to stress-test inputs}
\usage{
st_data_paths(
  data = Sys.getenv("ST_DATA_PATH"),
  project = Sys.getenv("ST_PROJECT_FOLDER")
)
}
\arguments{
\item{data}{Character. Path to the data directory.}

\item{project}{Character. Path to the project directory.}
}
\value{
A character vector.
}
\description{
This function helps expose the paths to \code{run_stress_test()} inputs.
}
\examples{
library(withr)

st_data_paths()

st_data_paths("/my/data", "/my/project/")

local({
  local_envvar(st_data_paths("/my/data", "/my/project/"))

  print(Sys.getenv("ST_DATA_PATH"))
  print(Sys.getenv("ST_PROJECT_FOLDER"))
})
Sys.getenv("ST_DATA_PATH")
Sys.getenv("ST_PROJECT_FOLDER")
}
\concept{helpers}
