% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_test_arguments.R
\docType{data}
\name{stress_test_arguments}
\alias{stress_test_arguments}
\title{A dataset describing the arguments of \code{\link[=run_stress_test]{run_stress_test()}}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11 rows and 6 columns.
}
\usage{
stress_test_arguments
}
\description{
A dataset describing the arguments of \code{\link[=run_stress_test]{run_stress_test()}}
}
\examples{
stress_test_arguments
}
\keyword{datasets}
