% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_test_arguments_combinations.R
\docType{data}
\name{stress_test_arguments_combinations}
\alias{stress_test_arguments_combinations}
\title{A dataset describing the combinations of arguments of \code{\link[=run_trisk]{run_trisk()}} and \code{\link[=run_lrisk]{run_lrisk()}}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 428 rows and 4 columns.
}
\usage{
stress_test_arguments_combinations
}
\description{
A dataset describing the combinations of arguments of \code{\link[=run_trisk]{run_trisk()}} and \code{\link[=run_lrisk]{run_lrisk()}}
}
\examples{
stress_test_arguments_combinations
}
\keyword{datasets}
