% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_test_arguments_numeric.R
\docType{data}
\name{stress_test_arguments_numeric}
\alias{stress_test_arguments_numeric}
\title{A dataset describing the arguments of \code{\link[=run_trisk]{run_trisk()}} and \code{\link[=run_lrisk]{run_lrisk()}}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13 rows and 6 columns.
}
\usage{
stress_test_arguments_numeric
}
\description{
A dataset describing the arguments of \code{\link[=run_trisk]{run_trisk()}} and \code{\link[=run_lrisk]{run_lrisk()}}
}
\examples{
stress_test_arguments_numeric
}
\keyword{datasets}
