% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{subtract_settlement}
\alias{subtract_settlement}
\title{Calculate annual profits after payout of settlement in lrisk}
\usage{
subtract_settlement(
  data,
  scc,
  exp_share_damages_paid,
  settlement_factor,
  shock_year
)
}
\arguments{
\item{data}{data frame containing the full trajectory company data}

\item{scc}{Numeric. Social cost of carbon per excess ton of CO2 emitted. This
is the price for each surplus ton of CO2 that goes into the calculation of
the carbon liability of a company.}

\item{exp_share_damages_paid}{Numeric. Ratio that defines the expected share
of the calculated social cost of carbon that is considered in the liability.}

\item{settlement_factor}{Catch all factor (ratio) that can be used to adjust
the expected payout of the settlement due to further data gaps. Set to 1 by
default.}

\item{shock_year}{Numeric, year of the litigation event}
}
\value{
A tibble holding annual profits
}
\description{
Calculate annual profits after payout of settlement in lrisk
}
