% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_data_has_expected_cols}
\alias{validate_data_has_expected_cols}
\title{Validate that a data frame contains expected columns}
\usage{
validate_data_has_expected_cols(data, expected_columns)
}
\arguments{
\item{data}{data frame that is to be validated}

\item{expected_columns}{Character vector listing the expected columns}
}
\value{
A boolean.
}
\description{
Validate that all expected columns for an operation are given in a data frame.
}
