% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_file_exists}
\alias{validate_file_exists}
\title{Validate that a file exists in a given directory}
\usage{
validate_file_exists(path)
}
\arguments{
\item{path}{Character vector indicating the directory of a file}
}
\value{
A boolean.
}
\description{
Before performing an operation on a file assumed to be found in a given
directory, validate this file exists and give indicative error if not.
}
