% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_input_values}
\alias{validate_input_values}
\title{Check that input values are valid}
\usage{
validate_input_values(
  lgd_senior_claims,
  lgd_subordinated_claims,
  risk_free_rate,
  discount_rate,
  div_netprofit_prop_coef,
  shock_year,
  fallback_term,
  use_company_terms,
  asset_type
)
}
\arguments{
\item{lgd_senior_claims}{Numeric, holding the loss given default for senior
claims, for accepted value range check \code{stress_test_arguments}.}

\item{lgd_subordinated_claims}{Numeric, holding the loss given default for
subordinated claims, for accepted value range check
\code{stress_test_arguments}.}

\item{risk_free_rate}{Numeric that indicates the risk free rate of interest.
For accepted range compare \code{stress_test_arguments}.}

\item{discount_rate}{Numeric, that holds the discount rate of dividends per
year in the DCF. For accepted range compare \code{stress_test_arguments}.}

\item{div_netprofit_prop_coef}{Numeric. A coefficient that determines how
strongly the future dividends propagate to the company value. For accepted
range compare \code{stress_test_arguments}.}

\item{shock_year}{Numeric, holding year the shock is applied. For accepted
range compare \code{stress_test_arguments}.}

\item{fallback_term}{Numeric. A coefficient that determines for which
maturity the expected loss should be calculated in the credit risk section
in case no company level term data are provided via \code{use_company_terms}.
For accepted range compare \code{stress_test_arguments}.}

\item{use_company_terms}{Boolean, indicating if term values for individual
companies are to be used. For accepted values compare
\code{stress_test_arguments}. Note that currently this functionality is not
available for asset_type bonds.}

\item{asset_type}{String holding asset_type. For accepted values compare
\code{stress_test_arguments}.}
}
\description{
Checks that user inputs are within defined ranges.
}
