% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_value_in_range}
\alias{validate_value_in_range}
\title{Validate that value in within range}
\usage{
validate_value_in_range(var, args_list)
}
\arguments{
\item{var}{String holding name of variable.}

\item{args_list}{Named list holding arguments of parent function call and
their values.}
}
\description{
Checks that value of variable \code{var} is in valid range as defined in
r2dii.climate_stress_test::stress_test_arguments.
}
