% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_values_in_range}
\alias{validate_values_in_range}
\title{Validate that values are within range}
\usage{
validate_values_in_range(var, args_list)
}
\description{
Checks that values of variable \code{var} are in valid range as defined in
r2dii.climate_stress_test::stress_test_arguments.
}
