% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_values_in_values}
\alias{validate_values_in_values}
\title{Validate that values are in within values}
\usage{
validate_values_in_values(var, args_list)
}
\arguments{
\item{var}{String holding name of variable.}

\item{args_list}{Named list holding arguments of parent function call and
their values.}
}
\description{
#' Checks that values of variable \code{var} are in valid values as defined in
r2dii.climate_stress_test::stress_test_arguments.
}
