% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_and_check.R
\name{wrangle_and_check_pacta_results}
\alias{wrangle_and_check_pacta_results}
\title{Wrangle and check PACTA results}
\usage{
wrangle_and_check_pacta_results(pacta_results, start_year, time_horizon)
}
\arguments{
\item{pacta_results}{Results from PACTA analysis.}
}
\value{
Wrangled \code{pacta_results.}
}
\description{
Function applies several filter steps to
\enumerate{
\item restrict PACTA results to scenarios useable for stresstesting
\item to apply run/project related selections.
}
}
\details{
Also consistency checks are run.
}
