% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_and_check.R
\name{wrangle_and_check_pacta_results_eq_cb}
\alias{wrangle_and_check_pacta_results_eq_cb}
\title{Wrangle and check PACTA results for asset types equity and bonds}
\usage{
wrangle_and_check_pacta_results_eq_cb(
  pacta_results,
  start_year,
  time_horizon,
  scenario_geography_filter,
  scenarios_filter,
  equity_market_filter
)
}
\arguments{
\item{pacta_results}{Results from PACTA analysis of type equity or bond.}

\item{start_year}{First considered year in analysis.}

\item{scenario_geography_filter}{Character. A vector of length 1 that
indicates which geographic scenario to apply in the analysis.}

\item{scenarios_filter}{Vector holding baseline and shock scenario name.}

\item{equity_market_filter}{Character. A vector of length 1 that
indicates which equity marjet to apply in the analysis.}
}
\value{
Wrangled \code{pacta_results.}
}
\description{
Function applies several filter steps to
\enumerate{
\item restrict PACTA results to scenarios useable for stresstesting
\item to apply run/project related selections.
}
}
\details{
Also consistency checks are run.
}
