% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_and_check.R
\name{wrangle_and_check_production_data}
\alias{wrangle_and_check_production_data}
\title{Wrangle and check production input data}
\usage{
wrangle_and_check_production_data(data, start_year, time_horizon)
}
\arguments{
\item{data}{A tibble holding production data.}

\item{start_year}{Numeric. Start year of the analysis}

\item{time_horizon}{Numeric. Number of years of forward looking production data}
}
\value{
Returns data invisibly.
}
\description{
Wrangle and check production input data
}
