% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_and_check.R
\name{wrangle_and_check_sector_exposures_eq_cb}
\alias{wrangle_and_check_sector_exposures_eq_cb}
\title{Wrangles and checks sector exposures}
\usage{
wrangle_and_check_sector_exposures_eq_cb(sector_exposures, asset_type)
}
\arguments{
\item{sector_exposures}{A tibble holding sector_exposures as calculated by
PACTA.}

\item{asset_type}{Type of analyzed asset. Can be bonds or equity.}
}
\value{
\code{sector_exposures} holding only valid rows for \code{asset_type}
}
\description{
Wrangles sector exposures by removing invalid rows and rows from assets other
than \code{asset_types}. Also checks for duplicates and that more than 0 rows
remain.
}
