% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_and_check.R
\name{wrangle_financial_data}
\alias{wrangle_financial_data}
\title{Wrangle financial data}
\usage{
wrangle_financial_data(financial_data, asset_type)
}
\arguments{
\item{financial_data}{A data set of \code{financial_data}.}

\item{asset_type}{A string indicating if company data are for analysis for
bond or equity.}
}
\value{
A prewrangled \code{financial_data} set.
}
\description{
Applies custom improvements of \code{net_profit_margin} and does row and cols
selections as well as filtering. Rows that have implausible
net_profit_margins below or equal to 0 are removed.
}
