% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_and_check.R
\name{wrangle_scenario_data}
\alias{wrangle_scenario_data}
\title{Wrangle scenario data}
\usage{
wrangle_scenario_data(scenario_data, start_year, end_year)
}
\arguments{
\item{scenario_data}{A tibble holding scenario data.}

\item{start_year}{Start year of analysis.}

\item{end_year}{End year of analysis.}
}
\value{
A tibble holding wrangled scenario_data.
}
\description{
Function applies custom wrangling to scenario data.
}
