% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_log}
\alias{write_log}
\title{Write error logs to the project log files. Appends the most recent message
to that file.}
\usage{
write_log(msg, location, ...)
}
\arguments{
\item{msg}{A string containing the error message}

\item{location}{A string containing the path to the project directory}

\item{...}{unnamed arguments are pasted at the end of the log message.}
}
\value{
A string.
}
\description{
Write error logs to the project log files. Appends the most recent message
to that file.
}
\seealso{
Other miscellaneous utility functions: 
\code{\link{data_path}()},
\code{\link{degrees}()},
\code{\link{dropbox_exists}()},
\code{\link{get_st_data_path}()}
}
\concept{miscellaneous utility functions}
