% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_results.R
\name{write_results}
\alias{write_results}
\title{Write the results to the project directory}
\usage{
write_results(
  data,
  path_to_results = NULL,
  investorname = NULL,
  asset_type = NULL,
  level = NULL,
  file_type = NULL
)
}
\arguments{
\item{data}{A dataframe the results of the stress test for one asset type}

\item{path_to_results}{Character. A string that contains the path that the
result should be written to.}

\item{investorname}{Character. A string containing the name of the investor
for which the analysis was run. Is used as part of the file name.}

\item{asset_type}{Character. A vector of length one indicating which type of
asset the results belong to. Currently supports "equity" and "bonds".}

\item{level}{Character. A vector of length one indicating which type of output
format to use. This depends on whether portfolio or company level pacta
data were used as input to the analysis. Defaults to the level indicated
in the work flow. Supports "company" and "portfolio".}

\item{file_type}{Character. A string containing the type of file that should
be written to the result path. Currently supports "csv" and "rda".}
}
\description{
Write the results to the project directory
}
\concept{output functions}
