% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_results.R
\name{write_results_new}
\alias{write_results_new}
\title{Write the results to the project directory}
\usage{
write_results_new(
  data,
  path_to_results = NULL,
  asset_type = NULL,
  level = NULL,
  file_type = NULL,
  sensitivity_analysis_vars,
  iter_var
)
}
\arguments{
\item{data}{A dataframe the results of the stress test for one asset type}

\item{path_to_results}{Character. A string that contains the path that the
result should be written to.}

\item{asset_type}{Character. A vector of length one indicating which type of
asset the results belong to. Currently supports "equity" and "bonds".}

\item{level}{Character. A vector of length one indicating which type of output
format to use. This depends on whether portfolio or company level pacta
data were used as input to the analysis. Defaults to the level indicated
in the work flow. Supports "company" and "portfolio".}

\item{file_type}{Character. A string containing the type of file that should
be written to the result path. Currently supports "csv" and "rda".}

\item{sensitivity_analysis_vars}{String vector holding names of iteration
arguments.}

\item{iter_var}{String holding name of iteration variable.}
}
\description{
Unlike \code{\link[=write_results]{write_results()}} as used in webscript function only allows level
company and does not create a new directory level by \code{investorname}.
}
\seealso{
Other output functions: 
\code{\link{write_results}()}
}
\concept{output functions}
