% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_results.R
\name{write_stress_test_results}
\alias{write_stress_test_results}
\title{Write stress test reports to output dir}
\usage{
write_stress_test_results(results_list, asset_type, iter_var, output_path)
}
\arguments{
\item{results_list}{A list of st results.}

\item{asset_type}{String holding asset type.}

\item{iter_var}{String holding name of iteration variable.}

\item{output_path}{String holding path to output dir.}
}
\description{
Stress test results are  exported to the output dir.
}
