% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_results.R
\name{aggregate_results}
\alias{aggregate_results}
\title{Aggregate results}
\usage{
aggregate_results(results_list, sensitivity_analysis_vars)
}
\arguments{
\item{results_list}{A list of results.}

\item{sensitivity_analysis_vars}{String vector holding names of iteration
arguments.}
}
\value{
A list of including basic and aggregated results.
}
\description{
Function aggregates results to the expected levels. List element entry
\code{results} is split into market risk results for company and portfolio level.
}
