% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_survival_probability_merton.R
\name{calc_survival_probability_merton}
\alias{calc_survival_probability_merton}
\title{Calculate survival probability}
\usage{
calc_survival_probability_merton(L, V0, sigma, r, t)
}
\arguments{
\item{L}{Numeric vector, holding debt values at maturity.}

\item{V0}{Numeric vector, holding company values at time t0.}

\item{sigma}{Numeric vector, holding volatility values.}

\item{r}{Numeric vector, holding risk free interest rates.}

\item{t}{Vector vector holding debt maturities.}
}
\value{
A vector holding survival probabilities,
}
\description{
Function calculates survival probability for a maturity based on a structural
Merton model.
For details on implementation please compare \code{\link[CreditRisk:Merton]{CreditRisk::Merton()}}.
Unlike \code{\link[CreditRisk:Merton]{CreditRisk::Merton()}} this implementation:
\enumerate{
\item only holds functionality to calculate probability of survival
\item can be called in vectorised fashion
\item has additional input validation by checking that all input values are
non negative numeric vectors of the same length.
}
}
