% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_exposure_by_technology_and_company}
\alias{calculate_exposure_by_technology_and_company}
\title{Calculate exposure_by_technology_and_company}
\usage{
calculate_exposure_by_technology_and_company(
  asset_type,
  input_data_list,
  start_year,
  time_horizon,
  scenario_to_follow_ls,
  log_path
)
}
\arguments{
\item{asset_type}{String holding type of asset.}

\item{input_data_list}{List with project agnostic and project specific input data}

\item{start_year}{Numeric, holding start year of analysis.}

\item{time_horizon}{Considered timeframe for PACTA analysis.}

\item{scenario_to_follow_ls}{Character. A string that indicates which
of the scenarios included in the analysis should be used to set the
late & sudden technology trajectories.}

\item{log_path}{String holding path to log file.}
}
\value{
A tibble holding exposure_by_technology_and_company.
}
\description{
Wrapper to calculate exposure_by_technology_and_company.
}
