% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_functions.R
\name{calculate_net_profits}
\alias{calculate_net_profits}
\title{Calculates annual net profits on the company-technology level for the
baseline and late and sudden scenarios. Climate laggards which need to build
out their production in increasing technologies to compensate for their
missed targets, are "punished" by applying their proximity to target as an
adjustment factor on the net profit margin. This ensures that late build out
will not simply translate into increased profits. Additionally, we adjust by
the market share of the company on the given technology. This accounts for
the fact that companies of different size will have different degrees of
difficulty catching up.}
\usage{
calculate_net_profits(data)
}
\arguments{
\item{data}{A data frame containing the production forecasts of companies
under baseline and late and sudden, market prices/costs, company net profit
margins, the proximity to target in the production forecast period and an
indication of the direction of the technology.}
}
\description{
Calculates annual net profits on the company-technology level for the
baseline and late and sudden scenarios. Climate laggards which need to build
out their production in increasing technologies to compensate for their
missed targets, are "punished" by applying their proximity to target as an
adjustment factor on the net profit margin. This ensures that late build out
will not simply translate into increased profits. Additionally, we adjust by
the market share of the company on the given technology. This accounts for
the fact that companies of different size will have different degrees of
difficulty catching up.
}
