% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_functions.R
\name{calculate_net_profits_shock_declining_technologies}
\alias{calculate_net_profits_shock_declining_technologies}
\title{Calculates annual net profits on the company-ald_business_unit level for the shock scenario for declining technologies}
\usage{
calculate_net_profits_shock_declining_technologies(data)
}
\arguments{
\item{data}{A data frame containing the production forecasts of companies
under baseline and late and sudden, market prices/costs, company net profit
margins, the proximity to target in the production forecast period and an
indication of the direction of the ald_business_unit.}
}
\value{
A data frame with net profits of companies with a declining ald_business_unit
}
\description{
Calculates annual net profits on the company-ald_business_unit level for the shock scenario for declining technologies
}
