% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_and_check.R
\name{check_company_terms}
\alias{check_company_terms}
\title{Check company term values for plausibility}
\usage{
check_company_terms(data)
}
\arguments{
\item{data}{A tibble holding company_terms data.}
}
\value{
Returns data invisibly.
}
\description{
Check company term values for plausibility
}
