% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_and_check.R
\name{check_financial_data}
\alias{check_financial_data}
\title{Check financial data}
\usage{
check_financial_data(financial_data, interactive_mode = FALSE)
}
\arguments{
\item{financial_data}{A data set of \code{financial_data}.}

\item{interactive_mode}{If TRUE the, more verbose, interactive mode is used.}
}
\value{
Returns prewrangled \code{financial_data} invisibly.
}
\description{
Applies sanity checks to financial data.
}
\examples{
fin_data <- tibble::tibble(
  company_name = c("Firm A", "Firm B"),
  company_id = c(1, 2),
  corporate_bond_ticker = c(NA, "TICK1"),
  pd = c(0.01, 0.002),
  net_profit_margin = c(0.423, 0.2),
  debt_equity_ratio = c(0.1, 0.201),
  volatility = c(0.130, 0.299)
)

check_financial_data(
  financial_data = fin_data
)
}
