% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_expected_loss.R
\name{company_expected_loss}
\alias{company_expected_loss}
\title{Calculate change of expected loss between scenarios for corporate loans on
the company-technology level. The function uses a maximum term of 5 years and
all loans with longer maturities are thrown into the 5 year bucket. The
calculation of PD changes that underlies the input to this function is based
on the comparison of overall NPVs of the corresponding equity values for the
companies at hand. That is, the PD changes take into account all discounted
future profits plus the terminal value.}
\usage{
company_expected_loss(
  data,
  loss_given_default = NULL,
  exposure_at_default = NULL,
  port_aum = NULL
)
}
\arguments{
\item{data}{A dataframe containing the (discounted) annual profits}

\item{loss_given_default}{A numeric vector of length one that indicates
the loss given default for the given asset type. Usually either a senior
claim or a subordinated claim.}

\item{exposure_at_default}{A dataframe that contains the share of the
portfolio value of each company-technology combination. Used to quantify
the impact of the company-tech level shock on higher levels of aggregation
in the portfolio}

\item{port_aum}{A dataframe containing the value of the portfolio for
the asset type at hand}
}
\description{
Calculate change of expected loss between scenarios for corporate loans on
the company-technology level. The function uses a maximum term of 5 years and
all loans with longer maturities are thrown into the 5 year bucket. The
calculation of PD changes that underlies the input to this function is based
on the comparison of overall NPVs of the corresponding equity values for the
companies at hand. That is, the PD changes take into account all discounted
future profits plus the terminal value.
}
