% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_technology_asset_value_at_risk.R
\name{company_technology_asset_value_at_risk}
\alias{company_technology_asset_value_at_risk}
\title{Calculate percentage value change between scenarios for equity (and
temporarily other asset types) on the company-ald_business_unit level}
\usage{
company_technology_asset_value_at_risk(
  data,
  shock_scenario = NULL,
  div_netprofit_prop_coef = NULL,
  flat_multiplier = NULL,
  crispy = FALSE
)
}
\arguments{
\item{data}{A dataframe containing the (discounted) annual profits}

\item{shock_scenario}{A dataframe containing the specification of the
shock scenario at hand}

\item{div_netprofit_prop_coef}{Numeric. A coefficient that determines how
strongly the future dividends propagate to the company value}

\item{flat_multiplier}{Numeric. A ratio that determines for the asset type
if how strongly the DCF should propagate to value changes.}

\item{crispy}{Boolean. Indicates if the output should be used for the CRISPY
database or for standard portfolio calculation (default).}
}
\description{
Calculate percentage value change between scenarios for equity (and
temporarily other asset types) on the company-ald_business_unit level
}
