% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_stressdata_masterdata_file_paths}
\alias{create_stressdata_masterdata_file_paths}
\title{Get path for stress test masterdata files}
\usage{
create_stressdata_masterdata_file_paths()
}
\value{
A list with 3 file paths.
}
\description{
Return list of paths for stress test masterdata files.
}
