% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{customise_output_path}
\alias{customise_output_path}
\title{Customise output path}
\usage{
customise_output_path(
  output_path,
  iter_var,
  shock_scenario,
  scenario_geography,
  carbon_price_model,
  financial_stimulus,
  risk_type
)
}
\arguments{
\item{output_path}{String holding path to st output folder.}

\item{iter_var}{String holding name of iteration variable.}

\item{shock_scenario}{String holding name of shock scenario.}

\item{scenario_geography}{String holding name of scenario geography.}

\item{carbon_price_model}{String holding name of carbon price model for trisk}

\item{risk_type}{String holding the risk type.}
}
\value{
Path to subdirectory in st output folder.
}
\description{
Checks for existence of provided \code{output_path} and extends it to hold a
subdirectory named after the timestamp and \code{iter_var}.
}
