% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_paths.R
\name{data_path}
\alias{data_path}
\title{conveniently access stress test related files in the directory set via envvar
or options}
\usage{
data_path(..., data_store = "data-raw")
}
\arguments{
\item{...}{Character vectors, if any values are \code{NA}, the result will also be
\code{NA}.}

\item{data_store}{String. Directory that contains the relevant data for the stress test analysis}
}
\value{
Character
}
\description{
conveniently access stress test related files in the directory set via envvar
or options
}
\examples{
data_path()
}
\seealso{
Other miscellaneous utility functions: 
\code{\link{degrees}()},
\code{\link{dropbox_exists}()},
\code{\link{get_st_data_path}()},
\code{\link{write_log}()}
}
\concept{miscellaneous utility functions}
