% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dropbox_exists}
\alias{dropbox_exists}
\title{Does this computer have a local copy of 2dii's dropbox folder?}
\usage{
dropbox_exists()
}
\value{
Logical.
}
\description{
Does this computer have a local copy of 2dii's dropbox folder?
}
\examples{
r2dii.climate.stress.test:::dropbox_exists()
}
\seealso{
Other miscellaneous utility functions: 
\code{\link{data_path}()},
\code{\link{degrees}()},
\code{\link{get_st_data_path}()}
}
\concept{miscellaneous utility functions}
