% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_loanbook_st.R
\name{format_loanbook_st}
\alias{format_loanbook_st}
\title{Transforms the PACTA for banks output to PACTA EQ/CB format to enable use
in stress test model.}
\usage{
format_loanbook_st(
  data,
  investor_name = "investor_name",
  portfolio_name = "portfolio_name",
  credit = loan_share_credit_type
)
}
\arguments{
\item{data}{A dataframe that contains augmented PACTA for banks results for
the project at hand. This dataframe must additionally contain information
on credit size (outstanding and limit).}

\item{investor_name}{Character. A string that indicates for which investor
the data should be transformed.}

\item{portfolio_name}{Character. A string that indicates for which portfolio
the data should be transformed. Must be a portfolio that belongs to the
indicated investor.}

\item{credit}{Character. Indicates whether the formatted output should return
data for outstanding credits or credit limits. The stress test will then be
run on the chosen indicator.}
}
\value{
data frame
}
\description{
Takes as input the results from a PACTA for banks run, augmented with credit
sizes (outstanding and limits), as provided by running the calc_loan_book.R
script. This data set is wrangled so that it fits the structure normally
produced by PACTA_analysis. Such a formatting is required in order to allow
seamless integration of PACTA for banks results (i.e. loan book alignment
data) with the climate transition risk stress test / scenario analysis.
}
\concept{utility functions}
