% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_companies.R
\name{generate_test_companies}
\alias{generate_test_companies}
\title{Generates a toy PACTA style portfolio with example companies}
\usage{
generate_test_companies()
}
\value{
data frame
}
\description{
Does not take any inputs. Generates a PACTA style portfolio that contains
companies representing stylized cases of interest for the transition risk
stress test.
}
\seealso{
Other example case functions: 
\code{\link{generate_test_capacity_factors}()},
\code{\link{generate_test_exposure}()},
\code{\link{generate_test_scenarios}()}
}
\concept{example case functions}
