% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_scenarios.R
\name{generate_test_scenarios}
\alias{generate_test_scenarios}
\title{Generates a toy scenario data set for all technologies}
\usage{
generate_test_scenarios()
}
\value{
data frame
}
\description{
Does not take any inputs. Generates a data frame that contains toy scenario
data in the format generally expected in the climate transition risk stress
test. This is done for all technologies.
}
\seealso{
Other example case functions: 
\code{\link{generate_price_data_wide}()},
\code{\link{generate_price_data}()},
\code{\link{generate_test_capacity_factors}()},
\code{\link{generate_test_companies}()},
\code{\link{generate_test_exclude_companies_data}()},
\code{\link{generate_test_exposure}()},
\code{\link{generate_test_financial_data}()}
}
\concept{example case functions}
