% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_st_data_path.R
\name{get_st_data_path}
\alias{get_st_data_path}
\title{Get the path to the required stress test data location. Can be provided
either via environment variable or options.}
\usage{
get_st_data_path(var = "ST_DATA_PATH")
}
\arguments{
\item{var}{String. Provide the name of the environment variable that points
to the data directory/repo. Default is "ST_DATA_PATH". If no environment
variable is provided and the default "ST_DATA_PATH" cannot be found, the
function tries to get the data path from options. In that case, provide
the name of the option that points to the data directory/repo. Default
option to get the path is, again, "ST_DATA_PATH".}
}
\value{
Character
}
\description{
Get the path to the required stress test data location. Can be provided
either via environment variable or options.
}
\seealso{
Other miscellaneous utility functions: 
\code{\link{data_path}()},
\code{\link{degrees}()},
\code{\link{dropbox_exists}()},
\code{\link{write_log}()}
}
\concept{miscellaneous utility functions}
