% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\docType{data}
\name{lookup}
\alias{lookup}
\alias{credit_type_lookup}
\alias{sectors_lookup}
\title{Lookup valid values}
\format{
An object of class \code{character} of length 2.

An object of class \code{character} of length 4.
}
\usage{
credit_type_lookup

sectors_lookup
}
\description{
Lookup valid values
}
\examples{

credit_type_lookup
sectors_lookup
}
\keyword{datasets}
