% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_prepare.R
\name{read_and_prepare_project_specific_data}
\alias{read_and_prepare_project_specific_data}
\title{Read and prepare project specific data}
\usage{
read_and_prepare_project_specific_data(
  asset_type,
  calculation_level,
  time_horizon,
  scenario_geography_filter,
  scenarios_filter,
  equity_market_filter,
  term,
  path
)
}
\arguments{
\item{asset_type}{String holding asset_type. For accepted values compare
\code{stress_test_arguments}.}

\item{calculation_level}{String holding level of calculation.}

\item{time_horizon}{Considered timefram for PACTA analysis.}

\item{scenario_geography_filter}{String holding geographic asset locations to
be considered in analysis.}

\item{scenarios_filter}{Vector holding baseline and shock scenario name.}

\item{equity_market_filter}{Character. A vector of length 1 that
indicates which equity market to apply in the analysis.}

\item{term}{Numeric. A coefficient that determines for which maturity the
expected loss should be calculated in the credit risk section. For accepted
range compare \code{stress_test_arguments}.}

\item{path}{String holding path to data.}
}
\value{
A list of lists holding prepared project specific data.
}
\description{
Function reads in data that are specific the project and conducts some
checking and wrangling. Also infers start_year of analysis.
}
