% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_capacity_factors.R
\name{read_capacity_factors}
\alias{read_capacity_factors}
\title{Read in power capacity factors from csv and check that all expected columns
are given.}
\usage{
read_capacity_factors(path = NULL)
}
\arguments{
\item{path}{A string that points to the location of the file containing the
capacity factors.}
}
\description{
Read in power capacity factors from csv and check that all expected columns
are given.
}
\seealso{
Other import functions: 
\code{\link{generate_transition_shocks}()},
\code{\link{read_capacity_factors_old}()},
\code{\link{read_financial_data}()},
\code{\link{read_ngfs_carbon_tax}()},
\code{\link{read_pacta_results}()},
\code{\link{read_transition_scenarios}()}
}
\concept{import functions}
