% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_capacity_factors.R
\name{read_capacity_factors}
\alias{read_capacity_factors}
\title{Read in power capacity factors from csv and check that all expected columns
are given.}
\usage{
read_capacity_factors(path = NULL, version = NULL)
}
\arguments{
\item{path}{A string that points to the location of the file containing the
capacity factors}

\item{version}{A string indicating whether to prepare the capacity factors
basedn on the old logic (2017 data set) or the new logic (long format,
multiple years and scenarios). Must be either "old" or "new".}
}
\description{
Read in power capacity factors from csv and check that all expected columns
are given.
}
\seealso{
Other import functions: 
\code{\link{read_pacta_results}()},
\code{\link{read_transition_scenarios}()}
}
\concept{import functions}
