% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_company_data.R
\name{read_company_data}
\alias{read_company_data}
\title{Read in company finacial data processed from eikon exports and AR master data
that contain information on multiple credit risk inputs and company
production plans, aggregated to the ticker/technology/year level for
\code{asset_type} bonds and aggregated to the company/technology/year level for
\code{asset_type} equity and aggregated to the loan_taker/technology/year level
for \code{àsset_type} loans}
\usage{
read_company_data(path = NULL, asset_type)
}
\arguments{
\item{path}{A string that points to the location of the file containing the
company financial data.}

\item{asset_type}{A string indicating if company data are for analysis for
bond or equity.}
}
\description{
Read in company finacial data processed from eikon exports and AR master data
that contain information on multiple credit risk inputs and company
production plans, aggregated to the ticker/technology/year level for
\code{asset_type} bonds and aggregated to the company/technology/year level for
\code{asset_type} equity and aggregated to the loan_taker/technology/year level
for \code{àsset_type} loans
}
\seealso{
Other import functions: 
\code{\link{generate_transition_shocks}()},
\code{\link{read_capacity_factors}()},
\code{\link{read_ngfs_carbon_tax}()},
\code{\link{read_pacta_results}()},
\code{\link{read_price_data}()},
\code{\link{read_transition_scenarios}()}
}
\concept{import functions}
