% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_company_terms.R
\name{read_company_terms}
\alias{read_company_terms}
\title{Read in company terms}
\usage{
read_company_terms(path)
}
\arguments{
\item{path}{Path to dir holding company terms data.}
}
\value{
A tibble holding company terms data.
}
\description{
Function reads in company terms and checks for existence of required columns.
}
