% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_financial_data.R
\name{read_financial_data}
\alias{read_financial_data}
\title{Read in company financial data processed from eikon exports and AR master data
that contain information on multiple credit risk inputs}
\usage{
read_financial_data(path = NULL)
}
\arguments{
\item{path}{A string that points to the location of the file containing the
company financial data.}
}
\description{
Read in company financial data processed from eikon exports and AR master data
that contain information on multiple credit risk inputs
}
\seealso{
Other import functions: 
\code{\link{read_capacity_factors_power}()},
\code{\link{read_carbon_data}()},
\code{\link{read_production_data}()}
}
\concept{import functions}
