% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ngfs_carbon_tax.R
\name{read_ngfs_carbon_tax}
\alias{read_ngfs_carbon_tax}
\title{Read in NGFS carbon tax data from csv and check that all expected columns
are given.}
\usage{
read_ngfs_carbon_tax(path = NULL)
}
\arguments{
\item{path}{A string that points to the location of the file containing the
NGFS carbon tax information}
}
\description{
Read in NGFS carbon tax data from csv and check that all expected columns
are given.
}
\seealso{
Other import functions: 
\code{\link{read_capacity_factors_old}()},
\code{\link{read_capacity_factors_power}()},
\code{\link{read_financial_data}()},
\code{\link{read_pacta_results_wt}()},
\code{\link{read_production_data}()},
\code{\link{read_transition_scenarios}()}
}
\concept{import functions}
