% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pacta_results.R
\name{read_pacta_results}
\alias{read_pacta_results}
\title{Read in the PACTA results from a given project and check that all expected
columns are given.}
\usage{
read_pacta_results(path = NULL)
}
\arguments{
\item{path}{A string that points to the location of the file containing the
pacta results}
}
\description{
Read in the PACTA results from a given project and check that all expected
columns are given.
}
\seealso{
Other import functions: 
\code{\link{read_capacity_factors_old}()},
\code{\link{read_capacity_factors}()},
\code{\link{read_financial_data}()},
\code{\link{read_pacta_results_wt}()},
\code{\link{read_transition_scenarios}()}
}
\concept{import functions}
