% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pacta_results.R
\name{read_pacta_results}
\alias{read_pacta_results}
\title{Read in the PACTA results from a given project and check that all expected
columns are given.}
\usage{
read_pacta_results(path = NULL, level = NULL)
}
\arguments{
\item{path}{A string that points to the location of the file containing the
pacta results}

\item{level}{A character vector of length 1 that indicates whether results on
the company level or on the portfolio level are to be analyzed.
Allowed values are "company", "portfolio".}
}
\description{
Read in the PACTA results from a given project and check that all expected
columns are given.
}
\seealso{
Other import functions: 
\code{\link{generate_transition_shocks}()},
\code{\link{read_capacity_factors}()},
\code{\link{read_financial_data}()},
\code{\link{read_ngfs_carbon_tax}()},
\code{\link{read_price_data}()},
\code{\link{read_transition_scenarios}()}
}
\concept{import functions}
