% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pacta_results.R
\name{read_pacta_results}
\alias{read_pacta_results}
\title{Read in power capacity factors from csv and check that all expected columns
are given.}
\usage{
read_pacta_results(path = NULL, asset_type = NULL, level = NULL)
}
\arguments{
\item{path}{A string that points to the location of the file containing the
capacity factors}

\item{asset_type}{A character vector of length 1 that indicates the type of
asset to be read in. Wrangling steps vary by asset, so this is essential.
Allowed values are "equity", "bonds", "loans".}

\item{level}{A character vector of length 1 that indicates whether results on
the company level or on the portfolio level are to be analyzed.
Allowed values are "company", "portfolio".}
}
\description{
Read in power capacity factors from csv and check that all expected columns
are given.
}
\seealso{
Other import functions: 
\code{\link{read_capacity_factors}()},
\code{\link{read_transition_scenarios}()}
}
\concept{import functions}
