% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_production_data.R
\name{read_production_data}
\alias{read_production_data}
\title{Read in company production data processed from AR PAMS data set that contains
information on production values and emission factors of companies}
\usage{
read_production_data(path = NULL)
}
\arguments{
\item{path}{A string that points to the location of the file containing the
company production data.}
}
\description{
Read in company production data processed from AR PAMS data set that contains
information on production values and emission factors of companies
}
\seealso{
Other import functions: 
\code{\link{read_capacity_factors_old}()},
\code{\link{read_capacity_factors_power}()},
\code{\link{read_financial_data}()},
\code{\link{read_ngfs_carbon_tax}()},
\code{\link{read_pacta_results_wt}()},
\code{\link{read_pacta_results}()},
\code{\link{read_transition_scenarios}()}
}
\concept{import functions}
