% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{report_duplicates}
\alias{report_duplicates}
\title{Report duplicate rows}
\usage{
report_duplicates(data, cols)
}
\arguments{
\item{data}{A tibble.}

\item{cols}{Cols to check for duplicate combinations on.}
}
\description{
Reports duplicates in \code{data} on columns \code{cols}.
}
