% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{report_missing_col_combinations}
\alias{report_missing_col_combinations}
\title{Identify and report missing value combinations}
\usage{
report_missing_col_combinations(data, composite_unique_cols, throw_error)
}
\arguments{
\item{data}{A tibble.}

\item{composite_unique_cols}{A vector of names of columns that shall be
unique in their combination.}

\item{throw_error}{Boolean, if TRUE error is thrown on failures, otherwise a
warning.}
}
\description{
Checks if all level combinations of \code{composite_unique_cols} are in\code{data} and
throws a warning on missing combinations.
NOTE:
\enumerate{
\item a combination of all levels is not neccesarily required/useful, make sure
to use function only in adequate context.
\item combiantions of too many columns/values may exceed memory size.
.
}
}
