% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_and_calc_loan_book.R
\name{run_prep_calculation_loans}
\alias{run_prep_calculation_loans}
\title{Run stress testing for loans}
\usage{
run_prep_calculation_loans(
  year_production_data,
  year_scenario_data,
  credit_type
)
}
\arguments{
\item{year_production_data}{Numeric. A vector of length 1 that indicates which
release year the production data should be taken from. For accepted
range compare \code{year_production_data_range_lookup}.}

\item{year_scenario_data}{Numeric. A vector of length 1 that indicates which
release year the scenario data should be taken from. For accepted
range compare \code{year_scenario_data_range_lookup}.}

\item{credit_type}{Type of credit. For accepted values please compare
\code{credit_type_loans}.}
}
\description{
Run stress testing for loans
}
