% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_prep_calculation_loans.R
\name{run_prep_calculation_loans}
\alias{run_prep_calculation_loans}
\title{Run stress testing for loans}
\usage{
run_prep_calculation_loans(credit_type = "outstanding")
}
\arguments{
\item{credit_type}{Type of credit. For accepted values please compare
\code{credit_type_lookup}.}
}
\description{
Run stress testing for loans
}
