% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_prep_calculation_loans.R
\name{run_prep_calculation_loans}
\alias{run_prep_calculation_loans}
\title{Run stress testing data preparation for loans}
\usage{
run_prep_calculation_loans(
  data_path_project_specific,
  data_path_project_agnostic,
  credit_type = "outstanding"
)
}
\arguments{
\item{data_path_project_specific}{String holding path to project specific
data.}

\item{data_path_project_agnostic}{String holding path to project agnostic
data.}

\item{credit_type}{Type of credit. For accepted values please compare
\code{credit_type_lookup}.}
}
\description{
Run stress testing data preparation for loans
}
