% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_and_check.R
\name{select_sector_scenario_combinations}
\alias{select_sector_scenario_combinations}
\title{Select required sector scenario combinations}
\usage{
select_sector_scenario_combinations(pacta_results)
}
\arguments{
\item{pacta_results}{A tibble holding pacta results.}
}
\value{
Tibble \code{pacta_results} holding only required sector - scenario
combinations.
}
\description{
Function: 1. Checks that for all supported sectors contained in
\code{pacta_results} required scenarios are available. 1. Filters so that only
required scenarios are kept.
}
