% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_bonds.R
\name{st_bonds}
\alias{st_bonds}
\alias{st_equity}
\alias{st_loans}
\alias{st_write_bonds}
\alias{st_write_equity}
\alias{st_write_loans}
\title{Sensitivity analysis: \code{run_stress_test()} multiple times}
\usage{
st_bonds(data = st_data_paths(), ..., quiet = TRUE)

st_equity(data = st_data_paths(), ..., quiet = TRUE)

st_loans(data = st_data_paths(), ..., quiet = TRUE)

st_write_bonds(data = st_data_paths(), destdir = tempdir(), ..., quiet = TRUE)

st_write_equity(data = st_data_paths(), destdir = tempdir(), ..., quiet = TRUE)

st_write_loans(data = st_data_paths(), destdir = tempdir(), ..., quiet = TRUE)
}
\arguments{
\item{data}{Character vector of length 2. Paths to the directories
ST_INPUTS_MASTER and ST_TESTING_\{aset-type\}, e.g. ST_TESTING_BONDS,
respectively. These directories must have the files expected by the
stress-test project. The easiest way to create this object is with
\code{\link[=st_data_paths]{st_data_paths()}}.}

\item{...}{Arguments passed to \code{\link[=run_stress_test]{run_stress_test()}}.}

\item{quiet}{Logical. Print non-condition messages to the console?}

\item{destdir}{Character. Path to a destination directory.}
}
\value{
A data frame.
}
\description{
These functions performs the so called "sensitivity analysis" for the asset
type in their name. They run \code{\link[=run_stress_test]{run_stress_test()}} multiple times, iterating
over multiple values of one and only one argument of \code{\link[=run_stress_test]{run_stress_test()}}.
}
\examples{
\dontshow{if (r2dii.climate.stress.test:::is_registered_dev("/home/mauro")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(readr, warn.conflicts = FALSE)

# Give the paths to your data explicitely
data <- st_data_paths(
  data = "/home/mauro/tmp/st/ST_INPUTS_MASTER",
  project = "/home/mauro/tmp/st/ST_TESTING_BONDS"
)
out <- suppressWarnings(
  st_bonds(data, term = c(1, 2))
)

# Or set these environment variables (e.g. in .Renviron), then omit `data`
data

# Here we omit `data` and make the console output more verbose
out <- st_bonds(term = c(1, 2), quiet = FALSE)

# The data frame output helps you quickly explore and manipulate your results
subset(out, st_name == "port" & arg_value == 2)

# If you run out of memory, instead write output sequentially to disk
path <- file.path(tempdir(), "st")
dir.create(path)
suppressWarnings(
  st_write_bonds(destdir = path, term = c(1, 2))
)
list.files(path)

# Later you may read all results at once
files <- list.files(path, full.names = TRUE)
results <- read_csv(files, id = "arg_val", show_col_types = FALSE)
results

# The column `arg_val` helps you identify each run
unique(
  subset(results, grepl("term___2", arg_val), select = 1:3)
)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=st_data_paths]{st_data_paths()}}.

Other main functions: 
\code{\link{run_stress_test}()}
}
\concept{main functions}
