% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun_st_bonds.R
\name{st_bonds}
\alias{st_bonds}
\alias{st_equity}
\alias{st_loans}
\title{Sensitivity analysis: \code{run_stress_test()} multiple times}
\usage{
st_bonds(data, ..., quiet = TRUE)

st_equity(data, ..., quiet = TRUE)

st_loans(data, ..., quiet = TRUE)
}
\arguments{
\item{data}{Character vector of length 2. Paths to the directories
ST_INPUTS_MASTER and ST_TESTING_\{aset-type\}, e.g. ST_TESTING_BONDS,
respectively. These directories must have the files expected by the
stress-test project. The easiest way to create this object is with
\code{\link[=st_data_paths]{st_data_paths()}}.}

\item{...}{Arguments passed to \code{\link[=run_stress_test]{run_stress_test()}}.}

\item{quiet}{Logical. Print non-condiiton messages to the console?}
}
\value{
A data frame.
}
\description{
These functions performs the so called "sensitivity analysis" for the asset
type in their name. They run \code{\link[=run_stress_test]{run_stress_test()}} multiple times, iterating
over multiple values of one and only one argument of \code{\link[=run_stress_test]{run_stress_test()}}.
}
\examples{
\dontshow{if (r2dii.climate.stress.test:::is_registered_dev()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data <- st_data_paths(
  data = "/home/mauro/tmp/st/ST_INPUTS_MASTER",
  project = "/home/mauro/tmp/st/ST_TESTING_BONDS"
)

out <- st_bonds(data, term = c(1, 2), shock_year = 2031)

# The data frame output helps you quickly explore and manipulate your results
subset(out, st_name == "port" & arg_value == 2)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=st_data_paths]{st_data_paths()}}.

Other main functions: 
\code{\link{run_stress_test}()}
}
\concept{main functions}
