% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{st_process_agnostic}
\alias{st_process_agnostic}
\title{Title}
\usage{
st_process_agnostic(
  data,
  scenario_geography,
  baseline_scenario,
  shock_scenario,
  sectors_and_technologies_list,
  start_year,
  carbon_price_model,
  end_year,
  log_path
)
}
\arguments{
\item{data}{trisk loaded data}
}
\value{
processed input trisk data
}
\description{
Title
}
