% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_input_values}
\alias{validate_input_values}
\title{Check that input values are valid}
\usage{
validate_input_values(
  baseline_scenario,
  shock_scenario,
  scenario_geography,
  lgd,
  risk_free_rate,
  discount_rate,
  growth_rate,
  div_netprofit_prop_coef,
  shock_year,
  settlement_factor,
  exp_share_damages_paid,
  scc,
  risk_type,
  carbon_price_model,
  market_passthrough,
  financial_stimulus
)
}
\arguments{
\item{baseline_scenario}{Holds the name of the baseline scenario to be used
in the stress test, for accepted value range check \code{stress_test_arguments}.}

\item{shock_scenario}{Holds the name of the shock scenario to be used in the
stress test, for accepted value range check \code{stress_test_arguments}.}

\item{scenario_geography}{Character vector, indicating which geographical
region(s) (concerning asset location) results shall be calculated for. For
accepted values compare \code{stress_test_arguments}.}

\item{lgd}{Numeric, holding the loss given default, for accepted value range
check \code{stress_test_arguments}.}

\item{risk_free_rate}{Numeric that indicates the risk free rate of interest.
For accepted range compare \code{stress_test_arguments}.}

\item{discount_rate}{Numeric, that holds the discount rate of dividends per
year in the DCF. For accepted range compare \code{stress_test_arguments}.}

\item{growth_rate}{Numeric, that holds the terminal growth rate of profits
beyond the final year in the DCF. For accepted range compare
\code{stress_test_arguments}.}

\item{div_netprofit_prop_coef}{Numeric. A coefficient that determines how
strongly the future dividends propagate to the company value. For accepted
range compare \code{stress_test_arguments}.}

\item{shock_year}{Numeric, holding year the shock is applied. For accepted
range compare \code{stress_test_arguments}.}

\item{settlement_factor}{Catch all factor (ratio) that can be used to adjust the
expected payout of the settlement due to further data gaps. Set to 1 by
default.}

\item{exp_share_damages_paid}{Numeric. Ratio that defines the expected share
of the calculated social cost of carbon that is considered in the liability.
This can follow historical precedents such as the Tobacco Master
Settlement that had a an expected share of 2.7\% of the damages paid.}

\item{scc}{Numeric. Social cost of carbon per excess ton of CO2 emitted. This
is the price for each surplus ton of CO2 that goes into the calculation of
the carbon liability of a company.}

\item{risk_type}{String that is either lrisk or trisk.}

\item{carbon_price_model}{Character vector, indicating which NGFS model is used in regards to
carbon prices. Default is no carbon tax.}

\item{market_passthrough}{Firm's ability to pass carbon tax onto the consumer.}

\item{financial_stimulus}{Additional channel through which the net profits of green
companies can be boosted under a shock scenario.}
}
\description{
Checks that user inputs are within defined ranges.
}
