% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_input_values}
\alias{validate_input_values}
\title{Check that input values are valid}
\usage{
validate_input_values(
  lgd_senior_claims,
  lgd_subordinated_claims,
  terminal_value,
  risk_free_rate,
  discount_rate,
  div_netprofit_prop_coef,
  term,
  company_exclusion,
  credit_type = NULL
)
}
\arguments{
\item{lgd_senior_claims}{Numeric, holding the loss given default for senior
claims, for accepted value range check \code{lgd_senior_claims_range_lookup}.}

\item{lgd_subordinated_claims}{Numeric, holding the loss given default for
subordinated claims, for accepted value range check
\code{lgd_subordinated_claims_range_lookup}.}

\item{terminal_value}{Numeric. A ratio to determine the share of the
discounted value used in the terminal value calculation beyond the
projected time frame. For accepted range compare \code{terminal_value_range_lookup}.}

\item{risk_free_rate}{Numeric that indicates the risk free rate of interest.
For accepted range compare \code{risk_free_rate_range_lookup}.}

\item{discount_rate}{Numeric, that holds the discount rate of dividends per
year in the DCF. For accepted range compare \code{discount_rate_range_lookup}.}

\item{div_netprofit_prop_coef}{Numeric. A coefficient that determines how
strongly the future dividends propagate to the company value. For accepted
range compare \code{div_netprofit_prop_coef_range_lookup}.}

\item{term}{Numeric. A coefficient that determines for which maturity the
expected loss should be calculated in the credit risk section. For accepted
range compare \code{term_range_lookup}.}

\item{company_exclusion}{Boolean, indicating if companies provided in dataset
excluded_companies.csv shall be excluded.}

\item{credit_type}{Type of credit. For accepted values please compare
\code{credit_type_loans}.}
}
\description{
Checks that user inputs are within defined ranges.
}
