% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_and_check.R
\name{wrangle_financial_data}
\alias{wrangle_financial_data}
\title{Wrangle financial data}
\usage{
wrangle_financial_data(financial_data, start_year)
}
\arguments{
\item{financial_data}{A data set of \code{financal_data}.}

\item{start_year}{String holding start year of analysis.}
}
\value{
A prewrangled \code{financal_data} set.
}
\description{
Applies custom improvements of \code{net_profit_margin} and does row and cols
selections as well as filtering.
}
